pars

Module basic

Generic parser and combinator building blocks.

This module provides the majority of the combinators that will be used to implement most parsers. Everything implemented in this module supports any combination of input and error types. When using pars, it is best to become familiar with the contents of this module, as the majority of parsing problems should have a solution implemented here.

Most functions that take a parser as a parameter and produce a transformed parser (A.K.A. combinators) are also available as methods on the Parse trait.

Macros

alt

Creates a parser that will match exactly one of its arguments.

permutation

Creates a parser that matches a permutation of a sequence of parsers.

seq

Creates a parser that matches a sequence of parsers.

Structs

RepeatedIter

An Iterator over repeated applications of a parser.

RepeatedUntilIter

An Iterator over repeated applications of a parser.

SeparatedIter

An Iterator over repeated applications of a parser separated by another parser.

Traits

Range

Trait representing a minimum and maximum number of repetitions.

Functions

array

Creates a parser that is repeated a constant number of times.

by_ref

Creates a parser out of a reference to another parser.

collect_repeated

Creates a parser that is repeated a number of times within a range.

collect_repeated_until

Creates a parser that is repeated until a sentinel is parsed.

collect_separated

Creates a parser that is applied repeatedly while interspersed with a separator.

complete

Creates a parser that fails if all input is not consumed.

cond

Creates a parser that produces a parsed value only if condition is true.

constant

Creates a parser that unconditionally produces a specific value.

delimited

Creates a parser that applies three parser and discards the parsed value of the first and last.

either

Creates a parser that attempts to parse with each of two parsers.

eof

A parser that succeeds on empty input, and fails otherwise.

err_into

Creates a parser whose parsing error is converted via Into.

error

Creates a parser that unconditionally produces an error.

flat_map

Maps the result of a parser onto a combinator to produce a new parser.

map

Creates a parser whose parsed result is transformed.

map_err

Creates a parser that transforms a parsing error.

map_res

Creates a parser whose result is transformed.

not

Creates a parser that succeeds only if parser fails.

ok_into

Creates a parser whose parsed result is converted via Into.

ok_try_into

Creates a parser whose parsed result is converted via TryInto.

opt

Creates a parser that attempts to parse with parser if possible.

or_default

Creates a parser that returns a fallback value instead of an error.

or_else

Creates a parser that returns a fallback value instead of an error.

or_value

Creates a parser that returns a fallback value instead of an error.

pair

Creates a parser that applies two parsers in sequence.

peek

Creates a parser that parses without consuming input.

pop

A parser that extracts one symbol from the input stream.

prefix

Creates a parser that applies two parsers and discards the first.

recognize

Creates a parser that returns a Span of parsed input.

remaining

Parser that returns the remaining input.

remaining_len

Parser that returns the length of the remaining input without consuming input.

repeated

Creates a parser that is repeated a number of times within a range.

repeated_until

Creates a parser that is repeated until a sentinel is parsed.

res_into

Creates a parser whose parsed result or parsing error are converted via Into.

res_try_into

Creates a parser whose parsed result is converted via TryInto.

separated

Creates a parser that is applied repeatedly while interspersed with a separator.

separated_pair

Creates a parser that applies three parsers and discards the second.

spanned

Creates a parser that returns a Span of parsed input in addition to the parsed value.

spanned_into

Creates a parser that also returns a span of the input that was parsed.

suffix

Creates a parser that applies two parsers and discards the second.

take

Creates a parser that takes a number of symbols from the input.

try_flat_map

Maps the result of a parser onto a combinator to produce a new parser.

try_map

Creates a parser whose parsed result is fallibly transformed.

try_repeated

Creates a parser that is repeated a number of times within a range.

try_repeated_until

Creates a parser that is repeated until a sentinel is parsed.

try_separated

Creates a parser that is applied repeatedly while interspersed with a separator.

verbatim

Creates a parser that attempts to match input verbatim.

verify

Creates a parser that succeeds only if the parsed value passes the given predicate.

with

Converts a parser to produce a different return value.

with_default

Converts a parser to produce a different return value, via Default.

with_value

Converts a parser to produce a different return value, via Clone.