pars::basic
Function collect_many

pub const fn collect_many<P, R, C, I>(
    parser: P,
    range: R,
) -> impl Parse<I, Parsed = C, Error = P::Error>
where
    P: Parse<I>,
    R: RangeBounds<usize>,
    I: Input,
    C: FromIterator<P::Parsed>,

Creates a parser that is repeated a number of times within a range.

collect_many produces a parser that will apply parser repeatedly. The number of times parser will be applied is defined by range. For example, if range is 3..=5, parser will be applied at least 3 times, and will be applied up to 5 times if possible. If parser fails before parsing 3 times (or whatever lower bound applies to range, if any), a parsing error is returned from the new parser. Once the upper bound is met (again, if any) parsing automatically ends successfully. parser returning a parsing error after meeting the lower bound also ends parsing for the new parser.

The values produced by parser are collected by the FromIterator trait implementation on the parsed type of the parser, which is usually deduced.

Note that the returned new parser does not allocate unless the FromIterator implementation allocates.

Also note that collect_many is a more generalized version of collect_many0, collect_many1 and collect_repeated. If range is .., the returned parser is equivalent to a collect_many0 parser. If range is 1.., the returned parser is equivalent to a collect_many1 parser. If range is count..=count, the returned parser is equivalent to a collect_repeated parser with a size of count.

See also Parse::collect_many.

Example

use pars::unicode::{strict::char_with_prop, prop::Alphabetic};

fn my_parser(input: &str) -> PResult<String, &str> {
    collect_many(char_with_prop(Alphabetic), 3..=5)
        .parse(input)
}

assert_eq!(my_parser.parse("hello world"), Ok(Success(String::from("hello"), " world")));
assert_eq!(my_parser.parse("hey there"), Ok(Success(String::from("hey"), " there")));
assert!(my_parser.parse("hi there").is_err());
assert_eq!(my_parser.parse("goodbye"), Ok(Success(String::from("goodb"), "ye")));