pars::basic
Function collect_many1

pub const fn collect_many1<P, C, I>(
    parser: P,
) -> impl Parse<I, Parsed = C, Error = P::Error>
where
    P: Parse<I>,
    I: Input,
    C: FromIterator<P::Parsed>,

Creates a parser that is repeated at east once.

collect_many1 produces a parser that will apply parser repeatedly until it returns a parsing error. The parsed values are collected via the FromIterator trait implementation on the parsed type of the new parser, which is generally deduced. If parser does not successfully parse at least once, a parsing error is returned.

Note that the returned new parser does not allocate unless the FromIterator implementation allocates.

See also Parse::collect_many1.

Example

use pars::unicode::{prop::Alphabetic, strict::char_with_prop};

fn word(input: &str) -> PResult<String, &str> {
    collect_many1(char_with_prop(Alphabetic)).parse(input)
}

assert_eq!(word.parse("hello world"), Ok(Success(String::from("hello"), " world")));
assert!(word.parse("").is_err());
assert!(word.parse(" ").is_err());