pars::basic

Function collect_separated

pub const fn collect_separated<P, S, R, C, I>(
    parser: P,
    range: R,
    separator: S,
) -> impl Parse<I, Parsed = C, Error = P::Error>
where
    P: Parse<I>,
    S: Parse<I, Error = P::Error>,
    R: Range,
    C: FromIterator<P::Parsed>,
    I: Input,

Creates a parser that is applied repeatedly while interspersed with a separator.

collect_separated produces a parser that will apply parser and separator repeatedly, alternatiging between the two. The parsing pattern starts with parser and ends with parser, such as parser separator parser. This is distinct from pair(parser, separator).repeated(...) as collect_separated will not parse a trailing separator, and a repeated pair must parse a trailing separator.

The provided range determines the minimum and maximum number of repatitions of parser. For example if range is 3..=5, parser must be able to parse at least 3 times, and up to 5 times. Similarly, .. corresponds to 0 or more repetitions, and 1.. corresponds to 1 or more. An exact number of repetitions can be specified with a single integer value such as 3 (but 3..=3 is also valid and equivalent). See Range for more information on ranges.

The values produced by parser are collected via the FromIterator trait on the, typically inferred, parsed value type of the new parser. The values produced by separator are discarded.

See also Parse::collect_separated.

Example

fn my_parser(input: &str) -> PResult<Vec<char>, &str> {
    // Note that `Vec<char>` is inferred from the return type of `my_parser`
    collect_separated(
        unicode::strict::char,
        ..,
        unicode::strict::verbatim(","),
    ).parse(input)
}

assert_eq!(
    my_parser.parse("a,b,c"),
    Ok(Success(vec!['a', 'b', 'c'], "")),
);

assert_eq!(
    my_parser.parse("a,b,c,"),
    Ok(Success(vec!['a', 'b', 'c'], ",")),
);