pars::basic

Function complete

pub const fn complete<P, I>(
    parser: P,
) -> impl Parse<I, Parsed = P::Parsed, Error = P::Error>
where
    P: Parse<I>,
    I: Input,

Creates a parser that fails if all input is not consumed.

complete produces a parser that applies parser and then verifies that all input was consumed. If there is still input remaining, an error is returned via Error::expected_eof.

See also Parse::complete.

Example

use pars::unicode::strict::verbatim;

fn my_parser(input: &str) -> PResult<(), &str> {
    complete(verbatim("hello").with(|| ())).parse(input)
}

assert_eq!(my_parser.parse("hello"), Ok(Success((), "")));
assert!(my_parser.parse("hello world").is_err());
assert!(my_parser.parse("hi").is_err());