pars::basic
Function many_until

pub const fn many_until<P, Q, F, R, I>(
    parser: P,
    sentinel_parser: Q,
    collect_fn: F,
) -> impl Parse<I, Parsed = (R, Q::Parsed), Error = P::Error>
where
    P: Parse<I>,
    Q: Parse<I, Error = P::Error>,
    F: for<'a> Fn(ManyUntilIter<'a, P, Q, I>) -> R,
    I: Input,

Creates a parser that is repeated until a sentinel is parsed.

many_until produces a parser that will apply parser repeatedly until sentinel_parser succeeds. That is, sentinel_parser is applied; if it fails, parser is applied before trying sentinel_parser again. If sentinel_parser succeeds, parsing is complete.

The values produced by parser are passed to collect_fn in the form of an iterator, ManyUntilIter. The value returned by collect_fn and the parsed value from sentinel_parser together as a 2-tuple become the parsed result of the overall new parser. If parser returns a parsing error at any point, a parsing error is returned from the overall new parser.

Note that the returned new parser does not allocate. Values produced by the iterator are obtained on demand by applying parser each time Iterator::next is called. Allocation will only occur if the user provided function collect_fn allocates to produce its result.

See also Parse::many_until.

Example

fn c_str_len(input: &[u8]) -> PResult<usize, &[u8]> {
    many_until(bytes::u8, bytes::verbatim(b"\x00"), |iter| iter.count())
        .map(|(len, _)| len) // we only want the length
        .parse(input)
}

assert!(c_str_len.parse(b"hello\x00world") == Ok(Success(5, b"world")));
assert!(c_str_len.parse(b"\x00") == Ok(Success(0, b"")));
assert!(c_str_len.parse(b"hello").is_err());