pars::basic

Function map

pub const fn map<P, F, R, I>(
    parser: P,
    map_fn: F,
) -> impl Parse<I, Parsed = R, Error = P::Error>
where
    P: Parse<I>,
    I: Input,
    F: Fn(P::Parsed) -> R,

Creates a parser whose parsed result is transformed.

The provided function, map_fn, is applied to the parsed result of parser if it parses successfully. The value returned from map_fn is the parsed result of the new parser.

If the the mapping operation could fail, use try_map instead.

See also Parse::map.

Example

fn inv_byte(input: &[u8]) -> PResult<u8, &[u8]> {
    map(bytes::u8, |byte| !byte).parse(input)
}

assert_eq!(inv_byte.parse(b"\x0f").unwrap().0, 0xf0);