pars::basic

Function or_else

pub const fn or_else<P, F, I>(
    parser: P,
    or_else_fn: F,
) -> impl Parse<I, Parsed = P::Parsed, Error = P::Error>
where
    P: Parse<I>,
    I: Input,
    F: Fn() -> P::Parsed,

Creates a parser that returns a fallback value instead of an error.

If parser fails to parse, or_else_fn is called and its return value is returned as a success value. The new parser will never return a parsing error.

See also Parse::or_else

Example

fn my_parser(input: &str) -> PResult<char, &str> {
    or_else(unicode::strict::char, || '\0').parse(input)
}

assert_eq!(my_parser.parse("hello"), Ok(Success('h', "ello")));
assert_eq!(my_parser.parse(""), Ok(Success('\0', "")));