pars::basic

Function prefix

pub const fn prefix<P, Q, I>(
    prefix: P,
    parser: Q,
) -> impl Parse<I, Parsed = Q::Parsed, Error = P::Error>
where
    P: Parse<I>,
    Q: Parse<I, Error = P::Error>,
    I: Input,

Creates a parser that applies two parsers and discards the first.

prefix produces a parser that applies prefix and parser in sequence. If both succeed, the parsed value returned by prefix is discarded and the parsed value returned by parser is returned by the new parser. If either parser fails, the corresponding error is returned from the new parser. When prefix fails, parser is not applied.

Example

fn my_parser(input: &str) -> PResult<char, &str> {
    prefix(
        unicode::strict::verbatim("hello"),
        unicode::strict::char,
    ).parse(input)
}

assert_eq!(my_parser.parse("hello world"), Ok(Success(' ', "world")));
assert!(my_parser.parse("hello").is_err());
assert!(my_parser.parse("goodbye world").is_err());