pars::basic

Function recognize

pub const fn recognize<P, I>(
    parser: P,
) -> impl Parse<I, Parsed = Span<I>, Error = P::Error>
where
    P: Parse<I>,
    I: Input,

Creates a parser that returns a Span of parsed input.

recognize produces a parser that applies parser and on success, it returns a Span of the input consumed by parser.

recognize is similar to spanned, except that the parsed value returned by parser is discarded.

See also Parse::recognize.

Example

fn my_parser(input: &[u8]) -> PResult<Span<&[u8]>, &[u8]> {
    recognize(bytes::be::u32).parse(input)
}

let res = my_parser.parse(b"\x01\x02\x03\x04\x05");
assert!(res.is_ok());
let Success(span, rem) = res.unwrap();
let span: &[u8] = span.into();
assert_eq!(span, b"\x01\x02\x03\x04");
assert_eq!(rem, b"\x05");