pars::basic

Function remaining

pub fn remaining<I: Input, E: Error<I>>(input: I) -> PResult<I, I, E>

Parser that returns the remaining input.

remaining is a parser that always succeeds and consumes all input. The parsed value is the entire input passed to the parser.

Example

fn my_parser(input: &[u8]) -> PResult<(u8, &[u8]), &[u8]> {
    bytes::u8.then(remaining).parse(input)
}

assert!(my_parser.parse(b"\x01\x02\x03\x04") == Ok(Success((1u8, &b"\x02\x03\x04"[..]), b"")));
assert!(my_parser.parse(b"\x01") == Ok(Success((1, &b""[..]), b"")));