pars::basic

Function remaining_len

pub fn remaining_len<I: Input, E: Error<I>>(input: I) -> PResult<usize, I, E>

Parser that returns the length of the remaining input without consuming input.

remaining_len is a parser that always succeeds and does not consume any input. The parsed value is the number of symbols in the input passed to the parser.

Example

fn my_parser(input: &[u8]) -> PResult<(u8, usize), &[u8]> {
    bytes::u8.then(remaining_len).parse(input)
}

assert!(my_parser.parse(b"\x01\x02\x03\x04") == Ok(Success((1u8, 3), b"\x02\x03\x04")));
assert!(my_parser.parse(b"\x01") == Ok(Success((1u8, 0), b"")));