pars::basic

Function res_into

pub const fn res_into<P, R, E, I>(
    parser: P,
) -> impl Parse<I, Parsed = R, Error = E>
where
    P: Parse<I>,
    I: Input,
    P::Parsed: Into<R>,
    P::Error: Into<E>,
    E: Error<I>,

Creates a parser whose parsed result or parsing error are converted via Into.

If parser parses successfully, its result value is converted to R using the Into<R> trait. If parser fails, the returned parsing error is converted to E using the Into<E> trait.

See also Parse::res_into.

Example

#[derive(Debug, PartialEq)]
struct MyError<'a>(&'a [u8]);

impl<'a> Error<&'a [u8]> for MyError<'a> { ... }
impl<'a> From<bytes::Error<&'a [u8]>> for MyError<'a> { ... }

fn my_parser(input: &[u8]) -> PResult<Vec<u8>, &[u8], MyError<'_>> {
    res_into(bytes::u8.array::<5>()).parse(input)
}

assert!(my_parser.parse(b"hello world") == Ok(Success(Vec::from(b"hello"), b" world")));
assert!(my_parser.parse(b"hi") == Err(Failure(MyError(b""), b"hi")));