pars::basic

Function separated_pair

pub const fn separated_pair<P, S, Q, I>(
    first: P,
    separator: S,
    second: Q,
) -> impl Parse<I, Parsed = (P::Parsed, Q::Parsed), Error = P::Error>
where
    P: Parse<I>,
    S: Parse<I, Error = P::Error>,
    Q: Parse<I, Error = P::Error>,
    I: Input,

Creates a parser that applies three parsers and discards the second.

separated_pair produces a parser that applies first, separator, and second in sequence. On Success, the parsed values returned by first and second are returned by the new parser as a tuple, and the parsed value returned by separator is discarded. If any of the parsers fails, the corresponding error is returned by the new parser and any remaining parsers are not applied.

Example

fn my_parser(input: &str) -> PResult<(char, char), &str> {
    separated_pair(
        unicode::strict::char,
        unicode::strict::verbatim(","),
        unicode::strict::char,
    ).parse(input)
}

assert_eq!(my_parser.parse("a,bcd"), Ok(Success(('a', 'b'), "cd")));
assert!(my_parser.parse("a,").is_err());
assert!(my_parser.parse(",b").is_err());