pars::basic
Function spanned

pub const fn spanned<P, I>(
    parser: P,
) -> impl Parse<I, Parsed = (P::Parsed, Span<I>), Error = P::Error>
where
    P: Parse<I>,
    I: Input,

Creates a parser that also returns a span of the input that was parsed.

spanned produces a parser that combines the parsed result of parser with a span of the input that parser parsed in a 2-tuple. This allows a parser to provide information about where in an input stream a parsed value originated from.

If the parsed value returned by parser is not needed, such as a parser that returns the unit type, consider using recognize instead.

See also Parse::spanned.

Example

fn my_parser(input: &[u8]) -> PResult<(u32, &[u8]), &[u8]> {
    spanned(bytes::be::u32)
        .map(|(val, span)| (val, span.into()))
        .parse(input)
}

assert!(my_parser.parse(b"\x01\x02\x03\x04\x05") ==
    Ok(Success((0x01020304, b"\x01\x02\x03\x04"), b"\x05")));
assert!(my_parser.parse(b"").is_err());