pars::basic
Function try_many

pub const fn try_many<P, R, F, O, S, I>(
    parser: P,
    range: R,
    collect_fn: F,
) -> impl Parse<I, Parsed = O, Error = P::Error>
where
    P: Parse<I>,
    R: RangeBounds<usize>,
    S: ErrorSeed<I, P::Error>,
    F: for<'a> Fn(ManyIter<'a, P, I>) -> Result<O, S>,
    I: Input,

Creates a parser that is repeated a number of times within a range.

try_many produces a parser that will apply parser repeatedly. The number of times parser will be applied is defined by range. For example, if range is 3..=5, parser will be applied at least 3 times, and will be applied up to 5 times if possible. If parser fails before parsing 3 times (or whatever lower bound applies to range, if any), a parsing error is returned from the new parser. Once the upper bound is met (again, if any) parsing automatically ends successfully. parser returning a parsing error after meeting the lower bound also ends parsing for the new parser.

The values produced by parser are passed to collect_fn in the form of an iterator, ManyIter. If collect_fn returns an Ok value, the contained value becomes the parsed value of the new parser. If collect_fn returns an Err, that gets returned as a parsing error via ErrorSeed.

Note that the returned new parser does not allocate. Values produced by the iterator are obtained on demand by applying parser each time Iterator::next is called. Allocation will only occur if the user provided function collect_fn allocates to produce its result.

Also note that try_many is a more generalized version of try_many0, try_many1 and try_repeated. If range is .., the returned parser is equivalent to a try_many0 parser. If range is 1.., the returned parser is equivalent to a try_many1 parser. If range is count..=count, the returned parser is equivalent to a [try_ repeated] parser with a size of count.

See also Parse::try_many.

Example

fn my_parser(input: &[u8]) -> PResult<u32, &[u8]> {
    try_many(bytes::u8, 2..=4, |iter| {
        let sum = iter.map(|byte| byte as u32).sum();
        // only allow even sums
        if sum % 2 == 0 {
            Ok(sum)
        } else {
            Err(ErrorKind::InvalidInput)
        }
    }).parse(input)
}

assert!(my_parser.parse(b"\x01\x02\x03\x04\x05") == Ok(Success(10, b"\x05")));
assert!(my_parser.parse(b"\x01\x02\x03") == Ok(Success(6, b"")));
assert!(my_parser.parse(b"\x01\x02").is_err());
assert!(my_parser.parse(b"\x01").is_err());
assert!(my_parser.parse(b"").is_err());