pars::basic
Function try_many0

pub const fn try_many0<P, F, R, S, I>(
    parser: P,
    collect_fn: F,
) -> impl Parse<I, Parsed = R, Error = P::Error>
where
    P: Parse<I>,
    I: Input,
    S: ErrorSeed<I, P::Error>,
    F: for<'a> Fn(Many0Iter<'a, P, I>) -> Result<R, S>,

Creates a parser that is repeated as many times as possible.

try_many0 produces a parser that will apply parser repeatedly until it returns a parsing error. The parsed values are passed to collect_fn in the form of an iterator, Many0Iter. If collect_fn returns an Ok value, that becomes the parsed result of the new parser. If collect_fn returns an Err, the new parser returns a parsing error by converting the contained value via ErrorSeed.

Because try_many0 permits any number of successful repetitions, including zero, the returned parser will only return a parsing error if collect_fn returns an Err.

Note that the returned new parser does not allocate. Values produced by the iterator are obtained on demand by applying parser each time Iterator::next is called. Allocation will only occur if the user provided function collect_fn allocates to produce its result.

See also Parse::try_many0.

Example

use pars::unicode::strict::verbatim;

// only allows an even number of spaces
fn count_spaces(input: &str) -> PResult<usize, &str> {
    try_many0(verbatim(" "), |iter| {
        let count = iter.count();
        if count % 2 == 0 {
            Ok(count)
        } else {
            Err(ErrorKind::InvalidInput)
        }
    }).parse(input)
}

assert_eq!(count_spaces.parse("    hello"), Ok(Success(4, "hello")));
assert_eq!(count_spaces.parse("hello"), Ok(Success(0, "hello")));
assert!(count_spaces.parse(" hello").is_err());