pars::basic

Function try_separated

pub const fn try_separated<P, S, R, F, E, O, I>(
    parser: P,
    range: R,
    separator: S,
    collect_fn: F,
) -> impl Parse<I, Parsed = O, Error = P::Error>
where
    P: Parse<I>,
    S: Parse<I, Error = P::Error>,
    R: Range,
    F: for<'a> Fn(SeparatedIter<'a, P, S, I>) -> Result<O, E>,
    E: ErrorSeed<I, P::Error>,
    I: Input,

Creates a parser that is applied repeatedly while interspersed with a separator.

try_separated produces a parser that will apply parser and separator repeatedly, alternatiging between the two. The parsing pattern starts with parser and ends with parser, such as parser separator parser. This is distinct from pair(parser, separator).try_repeated(...) as try_separated will not parse a trailing separator, and a repeated pair must parse a trailing separator.

The provided range determines the minimum and maximum number of repatitions of parser. For example if range is 3..=5, parser must be able to parse at least 3 times, and up to 5 times. Similarly, .. corresponds to 0 or more repetitions, and 1.. corresponds to 1 or more. An exact number of repetitions can be specified with a single integer value such as 3 (but 3..=3 is also valid and equivalent). See Range for more information on ranges.

The values produced by parser are passed to collect_fn in the form of an iterator, SeparatedIter. collect_fn then returns a Result, where the Ok variant becomes the parsed value of the new parser, and the Err variant becomes an error returned by the new parser via the ErrorSeed trait.

Note that the returned new parser does not allocate. Values produced by the iterator are obtained on demand by applying separator and parser each time Iterator::next is called. Allocation will only occur if the user provided collect_fn allocates when called.

See also Parse::try_separated.

Example

fn my_parser(input: &str) -> PResult<Vec<char>, &str> {
    try_separated(
        unicode::strict::char,
        ..,
        unicode::strict::verbatim(","),
        |iter| -> Result<Vec<char>, ErrorKind> {
            let mut out = Vec::new();
            for ch in iter {
                if !ch.is_ascii_alphabetic() {
                    return Err(ErrorKind::InvalidInput);
                }
                out.push(ch);
            }
            Ok(out)
        }
    ).parse(input)
}

assert_eq!(
    my_parser.parse("a,b,c"),
    Ok(Success(vec!['a', 'b', 'c'], "")),
);

assert_eq!(
    my_parser.parse("a,b,c,"),
    Ok(Success(vec!['a', 'b', 'c'], ",")),
);

assert!(my_parser.parse("a,2,c").is_err());