pars::basic

Function verbatim

pub const fn verbatim<P, I, E>(
    pattern: P,
) -> impl Parse<I, Parsed = Span<I>, Error = E>
where
    P: Input,
    I: Input,
    I::Symbol: PartialEq<P::Symbol>,
    E: Error<I>,

Creates a parser that attempts to match input verbatim.

pattern is a sequence of input symbols. The returned parser will succeed if the input begins with pattern exactly.

See also:

bytes::verbatim

ascii::strict::verbatim

ascii::lossy::verbatim

unicode::strict::verbatim

unicode::lossy::verbatim

Example

fn hello(input: &str) -> PResult<&str, &str> {
    verbatim("hello").ok_into().parse(input)
}

assert_eq!(hello.parse("hello world"), Ok(Success("hello", " world")));
assert!(hello.parse("Hello World").is_err());