pars::basic
Function with_default

pub const fn with_default<P, T, I>(
    parser: P,
) -> impl Parse<I, Parsed = T, Error = P::Error>
where
    P: Parse<I>,
    I: Input,
    T: Default,

Converts a parser to produce a different return value, via Default.

If parser parses successfully, its returned value is immediately dropped. The old returned value is replaced with a value of type T, produced by calling Default::default().

with_default(parser) is equivalent to with(parser, Default::default) (see with).

See also Parse::with_default.

Example

use pars::unicode::strict::verbatim;

#[derive(Debug, PartialEq, Default)]
struct Token;

fn token(input: &str) -> PResult<Token, &str> {
    with_default(verbatim("TOKEN")).parse(input)
}

assert_eq!(token.parse("TOKEN"), Ok(Success(Token, "")));