pars::basic
Function with_value

pub const fn with_value<P, T, I>(
    parser: P,
    value: T,
) -> impl Parse<I, Parsed = T, Error = P::Error>
where
    P: Parse<I>,
    T: Clone,
    I: Input,

Converts a parser to produce a different return value, via Clone.

If parser parses successfully, its returned value is immediately dropped. value is is cloned to replace the returned value.

See also Parse::with_value.

Example

use pars::unicode::strict::verbatim;

#[derive(Debug, PartialEq, Clone)]
struct Token;

fn token(input: &str) -> PResult<Token, &str> {
    with_value(verbatim("TOKEN"), Token).parse(input)
}

assert_eq!(token.parse("TOKEN"), Ok(Success(Token, "")));