pars::basic

Macro alt

alt!() { /* proc-macro */ }

Creates a parser that will match exactly one of its arguments.

alt will accept any number of parsers as arguments. The returned parser will attempt to parse with each argument, preferring the first argument, then second, and so on. Each alternate parser must produce the same output for a given type of input.

alt will attempt to parse each alternate completely before moving on to the next alternate. Thus, alt introduces backtracking into the parser. Depending on how each alternate parses, alt can become quite expensive, since each alternate may parse all the way to the end of input before returning a parse failure. That said, most use cases are not likely to have this worst case performance.

When backtracking is unacceptable, flat_map or try_flat_map may be a viable alternative. See the examples msgpack.rs and msgpack_alt.rs for a comparison of using alt versus try_flat_map.

Also see either, which alt is essentially a variadic version of, and in fact, is implemented in terms of.

Example

use pars::{
    basic::alt,
    unicode::{UnicodeInput as UInput, PResult, strict::verbatim},
    Parse, Success,
};

fn my_parser<I: UInput>(input: I) -> PResult<(), I> {
    // match "a", "b", or "c"
    alt!(
        verbatim("a").with_value(()),
        verbatim("b").with_value(()),
        verbatim("c").with_value(()),
    )
    .parse(input)
}

assert!(my_parser.parse("be kind") == Ok(Success((), "e kind")));