pars::basic

Trait Range

pub trait Range {
    // Required methods
    fn minimum(&self) -> usize;
    fn maximum(&self) -> usize;
}

Trait representing a minimum and maximum number of repetitions.

Range is used by combinators such as repeated and separated to specify either or both of the minimum and maximum number of repetitions. Range is implemented for built-in range values using the .. and ..= operators over usize values, as well as for plain usize values for an exact number of repetitions.

In parsing, repeating a pattern zero or more times, or one or more times are most common. .. represents zero or more, and 1.. represents one or more.

Example

let range = ..;
assert_eq!(range.minimum(), 0);
assert_eq!(range.maximum(), usize::MAX);

let range = ..10;
assert_eq!(range.minimum(), 0);
assert_eq!(range.maximum(), 9);

let range = 2..;
assert_eq!(range.minimum(), 2);
assert_eq!(range.maximum(), usize::MAX);

let range = 2..10;
assert_eq!(range.minimum(), 2);
assert_eq!(range.maximum(), 9);

let range = ..=10;
assert_eq!(range.minimum(), 0);
assert_eq!(range.maximum(), 10);

let range = 2..=10;
assert_eq!(range.minimum(), 2);
assert_eq!(range.maximum(), 10);

let range = 5;
assert_eq!(range.minimum(), 5);
assert_eq!(range.maximum(), 5);

Required Methods

fn minimum(&self) -> usize

The minimum number of repetitions

fn maximum(&self) -> usize

The maximum number of repetitions