pars
Struct ParsedInput

pub struct ParsedInput<P, I>
where
    P: Parse<I> + Clone,
    I: Input,
{ /* private fields */ }

Input derived from other input by repeatedly applying a parser.

ParsedInput consists of an underlying input and a parser. Each symbol produced by a ParsedInput is produced by applying the parser to the inner input. After parsing a symbol, the inner input becomes the remaining unparsed inner input. The ParsedInput ends when parsing a symbol fails.

Implementations

impl<P, I> ParsedInput<P, I>
where P: Parse<I> + Clone, I: Input,

pub const fn new(parser: P, input: I) -> Self

Creates a new ParsedInput.

The returned ParsedInput will produce symbols by repeated applying the provided parser over the provided input.