pars
Trait Input

pub trait Input: Clone {
    type Symbol;

    // Required methods
    fn next(&mut self) -> Option<Self::Symbol>;
    fn pos_eq(&self, other: &Self) -> bool;

    // Provided methods
    fn advance_by(&mut self, count: usize) -> usize { ... }
    fn size_hint(&self) -> Option<usize> { ... }
    fn is_empty(&self) -> bool { ... }
    fn map<F, O>(self, map_fn: F) -> impl Input<Symbol = O>
       where F: Clone + Fn(Self::Symbol) -> O { ... }
}

A parsable symbol stream.

Input is very similar to Iterator. The key difference is that a type implementing Input must support multiple and partial passes in any combination. pars implements recursive descent parsing, which means that there is no limit to the number of symbols the parser may consume before determining that it needs to backtrack (although individual parsers may have a finite bound).

As such, Input is essentially an Iterator that can be cloned and can be compared with clones for positional equality.

Required Associated Types

type Symbol

The symbol type this input provides a stream of.

Required Methods

fn next(&mut self) -> Option<Self::Symbol>

Gets the next symbol in the stream.

None indicates that the end of the input has been reached.

fn pos_eq(&self, other: &Self) -> bool

Compares the position of two inputs for positional equality.

If both inputs are at the same position of the stream, true is returned, false otherwise.

Provided Methods

fn advance_by(&mut self, count: usize) -> usize

Advances the stream by a given number of symbols.

The number of skipped symbols is returned, which should always be equal to count, unless the remaining input was less than count symbols. If the returned value is not equal to count, the input must now be at the end of input.

Implementors may wish to customize this method to provide a more performant implementation than calling next count times.

fn size_hint(&self) -> Option<usize>

Returns the remaining number of symbols, if known.

Not all input streams can easily know how many symbols they contain, but when it is known, some parsers can use that information for optimization.

It is recommended that implementors customize this method if the remaining size of the stream can be computed in constant time.

fn is_empty(&self) -> bool

Returns true if the stream is at the end of input.

Implementors may wish to customize this method to provide a more performant implementation. The default implementation clones the input and checks if calling next yeilds a value.

fn map<F, O>(self, map_fn: F) -> impl Input<Symbol = O>
where F: Clone + Fn(Self::Symbol) -> O,

Creates a new input stream with mapped values.

The provided function or closure will be used by the returned stream to transform symbols of the original stream.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.