pars

Trait Parse

pub trait Parse<I: Input> {
    type Parsed: Sized;
    type Error: Error<I>;

    // Required method
    fn parse<N>(&self, input: N) -> PResult<Self::Parsed, I, Self::Error>
       where N: IntoInput<Input = I>;

    // Provided methods
    fn map<F, R>(
        self,
        map_fn: F,
    ) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where Self: Sized,
             F: Fn(Self::Parsed) -> R { ... }
    fn try_map<F, R, S>(
        self,
        map_fn: F,
    ) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where Self: Sized,
             S: ErrorSeed<I, Self::Error>,
             F: Fn(Self::Parsed) -> Result<R, S> { ... }
    fn map_err<F, R>(
        self,
        map_err_fn: F,
    ) -> impl Parse<I, Parsed = Self::Parsed, Error = R>
       where Self: Sized,
             F: Fn(Self::Error) -> R,
             R: Error<I> { ... }
    fn map_res<F, R, E>(
        self,
        map_res_fn: F,
    ) -> impl Parse<I, Parsed = R, Error = E>
       where Self: Sized,
             F: Fn(Result<Self::Parsed, Self::Error>) -> Result<R, E>,
             E: Error<I> { ... }
    fn res_into<R, E>(self) -> impl Parse<I, Parsed = R, Error = E>
       where Self: Sized,
             Self::Parsed: Into<R>,
             Self::Error: Into<E>,
             E: Error<I> { ... }
    fn ok_into<R>(self) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where Self: Sized,
             Self::Parsed: Into<R> { ... }
    fn err_into<E>(self) -> impl Parse<I, Parsed = Self::Parsed, Error = E>
       where Self: Sized,
             Self::Error: Into<E>,
             E: Error<I> { ... }
    fn res_try_into<R, E>(self) -> impl Parse<I, Parsed = R, Error = E>
       where Self: Sized,
             Self::Parsed: TryInto<R>,
             Self::Error: Into<E>,
             E: Error<I> { ... }
    fn ok_try_into<R>(self) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where Self: Sized,
             Self::Parsed: TryInto<R> { ... }
    fn with<F, T>(
        self,
        with_fn: F,
    ) -> impl Parse<I, Parsed = T, Error = Self::Error>
       where Self: Sized,
             F: Fn() -> T { ... }
    fn with_default<T>(self) -> impl Parse<I, Parsed = T, Error = Self::Error>
       where Self: Sized,
             T: Default { ... }
    fn with_value<T>(
        self,
        value: T,
    ) -> impl Parse<I, Parsed = T, Error = Self::Error>
       where Self: Sized,
             T: Clone { ... }
    fn flat_map<F, P>(
        self,
        combinator: F,
    ) -> impl Parse<I, Parsed = P::Parsed, Error = Self::Error>
       where Self: Sized,
             F: Fn(Self::Parsed) -> P,
             P: Parse<I, Error = Self::Error> { ... }
    fn try_flat_map<F, P, S>(
        self,
        combinator: F,
    ) -> impl Parse<I, Parsed = P::Parsed, Error = Self::Error>
       where Self: Sized,
             S: ErrorSeed<I, Self::Error>,
             F: Fn(Self::Parsed) -> Result<P, S>,
             P: Parse<I, Error = Self::Error> { ... }
    fn or_value(
        self,
        value: Self::Parsed,
    ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
       where Self: Sized,
             Self::Parsed: Clone { ... }
    fn or_default(
        self,
    ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
       where Self: Sized,
             Self::Parsed: Default { ... }
    fn or_else<F>(
        self,
        or_else_fn: F,
    ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
       where Self: Sized,
             F: Fn() -> Self::Parsed { ... }
    fn repeated_until<P, F, R>(
        self,
        sentinel: P,
        collect_fn: F,
    ) -> impl Parse<I, Parsed = (R, P::Parsed), Error = Self::Error>
       where Self: Sized,
             P: Parse<I, Error = Self::Error>,
             F: for<'a> Fn(RepeatedUntilIter<'a, Self, P, I>) -> R { ... }
    fn try_repeated_until<P, F, R, S>(
        self,
        sentinel: P,
        collect_fn: F,
    ) -> impl Parse<I, Parsed = (R, P::Parsed), Error = Self::Error>
       where Self: Sized,
             S: ErrorSeed<I, Self::Error>,
             P: Parse<I, Error = Self::Error>,
             F: for<'a> Fn(RepeatedUntilIter<'a, Self, P, I>) -> Result<R, S> { ... }
    fn collect_repeated_until<P, R>(
        self,
        sentinel: P,
    ) -> impl Parse<I, Parsed = (R, P::Parsed), Error = Self::Error>
       where Self: Sized,
             P: Parse<I, Error = Self::Error>,
             R: FromIterator<Self::Parsed> { ... }
    fn repeated<N, F, R>(
        self,
        range: N,
        collect_fn: F,
    ) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where Self: Sized,
             N: Range,
             F: for<'a> Fn(RepeatedIter<'a, Self, I>) -> R { ... }
    fn try_repeated<N, F, R, S>(
        self,
        range: N,
        collect_fn: F,
    ) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where Self: Sized,
             N: Range,
             S: ErrorSeed<I, Self::Error>,
             F: for<'a> Fn(RepeatedIter<'a, Self, I>) -> Result<R, S> { ... }
    fn collect_repeated<C>(
        self,
        range: impl Range,
    ) -> impl Parse<I, Parsed = C, Error = Self::Error>
       where Self: Sized,
             C: FromIterator<Self::Parsed> { ... }
    fn array<const LEN: usize>(
        self,
    ) -> impl Parse<I, Parsed = [Self::Parsed; LEN], Error = Self::Error>
       where Self: Sized { ... }
    fn separated<N, S, F, R>(
        self,
        range: N,
        separator: S,
        collect_fn: F,
    ) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where Self: Sized,
             N: Range,
             S: Parse<I, Error = Self::Error>,
             F: for<'a> Fn(SeparatedIter<'a, Self, S, I>) -> R { ... }
    fn try_separated<N, S, F, R, E>(
        self,
        range: N,
        separator: S,
        collect_fn: F,
    ) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where Self: Sized,
             N: Range,
             S: Parse<I, Error = Self::Error>,
             F: for<'a> Fn(SeparatedIter<'a, Self, S, I>) -> Result<R, E>,
             E: ErrorSeed<I, Self::Error> { ... }
    fn collect_separated<C>(
        self,
        range: impl Range,
        separator: impl Parse<I, Error = Self::Error>,
    ) -> impl Parse<I, Parsed = C, Error = Self::Error>
       where Self: Sized,
             C: FromIterator<Self::Parsed> { ... }
    fn complete(
        self,
    ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
       where Self: Sized { ... }
    fn spanned(
        self,
    ) -> impl Parse<I, Parsed = (Self::Parsed, Span<I>), Error = Self::Error>
       where Self: Sized { ... }
    fn spanned_into<S>(
        self,
    ) -> impl Parse<I, Parsed = (Self::Parsed, S), Error = Self::Error>
       where Self: Sized,
             Span<I>: Into<S> { ... }
    fn recognize(self) -> impl Parse<I, Parsed = Span<I>, Error = Self::Error>
       where Self: Sized { ... }
    fn cond(
        self,
        condition: bool,
    ) -> impl Parse<I, Parsed = Option<Self::Parsed>, Error = Self::Error>
       where Self: Sized { ... }
    fn verify<F>(
        self,
        verify_fn: F,
    ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
       where Self: Sized,
             F: Fn(&Self::Parsed) -> bool { ... }
    fn not(self) -> impl Parse<I, Parsed = (), Error = Self::Error>
       where Self: Sized { ... }
    fn opt(
        self,
    ) -> impl Parse<I, Parsed = Option<Self::Parsed>, Error = Self::Error>
       where Self: Sized { ... }
    fn peek(self) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
       where Self: Sized { ... }
    fn by_ref<'a>(
        &'a self,
    ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error> + 'a
       where I: 'a { ... }
    fn then<P>(
        self,
        other: P,
    ) -> impl Parse<I, Parsed = (Self::Parsed, P::Parsed), Error = Self::Error>
       where Self: Sized,
             P: Parse<I, Error = Self::Error> { ... }
    fn or<P>(
        self,
        other: P,
    ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
       where Self: Sized,
             P: Parse<I, Parsed = Self::Parsed, Error = Self::Error> { ... }

}

Trait implemented by all parsers.

The Parse trait is where the logic of parsing is implemented. However, most pars users will not implement Parse directly. See the crate level documentation for examples of user defined parsers.

Required Associated Types

type Parsed: Sized

The value type that is produced by the parser on success.

type Error: Error<I>

The error type that is produced by the parser on failure.

Required Methods

fn parse<N>(&self, input: N) -> PResult<Self::Parsed, I, Self::Error>
where N: IntoInput<Input = I>,

Parses the provided input.

Provided Methods

fn map<F, R>(self, map_fn: F) -> impl Parse<I, Parsed = R, Error = Self::Error>
where Self: Sized, F: Fn(Self::Parsed) -> R,

Creates a parser whose parsed result is transformed.

The provided function, map_fn, is applied to the parsed result of self if it parses successfully. The value returned from map_fn is the parsed result of the new parser.

If the the mapping operation could fail, use Parse::try_map instead.

See also basic::map.

Example

fn inv_byte(input: &[u8]) -> PResult<u8, &[u8]> {
    bytes::u8.map(|byte| !byte).parse(input)
}

assert_eq!(inv_byte.parse(b"\x0f").unwrap().0, 0xf0);

fn try_map<F, R, S>( self, map_fn: F, ) -> impl Parse<I, Parsed = R, Error = Self::Error>
where Self: Sized, S: ErrorSeed<I, Self::Error>, F: Fn(Self::Parsed) -> Result<R, S>,

Creates a parser whose parsed result is fallibly transformed.

The provided function, map_fn, is applied to the parsed result of self if it parses successfully. map_fn is permitted to fail, which is signaled with an Err return value. In thise case the new parser fails, and the contained error is converted to an Error via ErrorSeed. If an Ok value is returned, the contained value becomes the parsed result of the new parser.

If the mapping function cannot fail, use Parse::map instead.

See also basic::try_map.

Example

fn inv_byte(input: &[u8]) -> PResult<u8, &[u8]> {
    bytes::u8.try_map(|byte| {
        if byte < 0x80 {
            Ok(!byte)
        } else {
            Err(ErrorKind::InvalidInput)
        }
    }).parse(input)
}

assert_eq!(inv_byte.parse(b"\x0f").unwrap().0, 0xf0);
assert!(inv_byte.parse(b"\xf0").is_err());

fn map_err<F, R>( self, map_err_fn: F, ) -> impl Parse<I, Parsed = Self::Parsed, Error = R>
where Self: Sized, F: Fn(Self::Error) -> R, R: Error<I>,

Creates a parser that transforms a parsing error.

The provided function, map_err_fn, is applied the parsing error of self if its parsing fails. The value returned from map_err_fn becomes the parsing error of the new parser. Otherwise, if self succeeds, its parsed result is the parsed result of the new parser.

This combinator is most useful for converting between error types.

See also basic::map_err.

Example

fn my_parser(input: &[u8]) -> PResult<u8, &[u8], unicode::Error<&[u8]>> {
    bytes::u8.map_err(|err: bytes::Error<&[u8]>| {
        match err.kind() {
            bytes::ErrorKind::NeedMoreInput => unicode::ErrorKind::NeedMoreInput,
            bytes::ErrorKind::ExpectedEof => unicode::ErrorKind::ExpectedEof,
            _ => unicode::ErrorKind::InvalidInput,
        }.into_error(*err.position())
    }).parse(input)
}

assert!(my_parser.parse(b"hello") == Ok(Success(b'h', b"ello")));
assert!(my_parser.parse(b"") == Err(Failure(
    unicode::ErrorKind::NeedMoreInput.into_error(b"".as_slice()), b"")));

fn map_res<F, R, E>(self, map_res_fn: F) -> impl Parse<I, Parsed = R, Error = E>
where Self: Sized, F: Fn(Result<Self::Parsed, Self::Error>) -> Result<R, E>, E: Error<I>,

Creates a parser whose result is transformed.

The provided function, map_res_fn, accepts a Result containing either the successfully parsed value of self or the parsing error returned by self. map_res_fn then returns a new Result, which will contain either the new successful parsed value or a new parsing error.

Unlike Parse::map or Parse::map_err, the mapping function may choose to convert an Ok into and Err or an Err into an Ok.

See also basic::map_res.

Example

fn my_parser(input: &[u8]) -> PResult<u8, &[u8]> {
    bytes::u8.map_res(|res| match res {
        Ok(byte) if byte >= 0x80 => {
            Err(ErrorKind::InvalidInput.into_error(b"".as_slice()))
        },
        Ok(byte) => Ok(byte),
        Err(err) if err.kind() == ErrorKind::NeedMoreInput => Ok(0),
        Err(err) => Err(err),
    }).parse(input)
}

assert!(my_parser.parse(b"hello") == Ok(Success(b'h', b"ello")));
assert_eq!(
    my_parser.parse(b"\xffhello").unwrap_err().0.kind(),
    ErrorKind::InvalidInput,
);
assert!(my_parser.parse(b"") == Ok(Success(b'\x00', b"")));

fn res_into<R, E>(self) -> impl Parse<I, Parsed = R, Error = E>
where Self: Sized, Self::Parsed: Into<R>, Self::Error: Into<E>, E: Error<I>,

Creates a parser whose parsed result or parsing error are converted via Into.

If self parses successfully, its result value is converted to R using the Into<R> trait. If self fails, the returned parsing error is converted to E using the Into<E> trait.

See also basic::res_into.

Example

#[derive(Debug, PartialEq)]
struct MyError<'a>(&'a [u8]);

impl<'a> Error<&'a [u8]> for MyError<'a> { ... }
impl<'a> From<bytes::Error<&'a [u8]>> for MyError<'a> { ... }

fn my_parser(input: &[u8]) -> PResult<Vec<u8>, &[u8], MyError<'_>> {
    bytes::u8.array::<5>().res_into().parse(input)
}

assert!(my_parser.parse(b"hello world") == Ok(Success(Vec::from(b"hello"), b" world")));
assert!(my_parser.parse(b"hi") == Err(Failure(MyError(b""), b"hi")));

fn ok_into<R>(self) -> impl Parse<I, Parsed = R, Error = Self::Error>
where Self: Sized, Self::Parsed: Into<R>,

Creates a parser whose parsed result is converted via Into.

If self parses successfully, its result value is converted to R using the Into<R> trait.

See also basic::ok_into.

Example

fn my_parser(input: &[u8]) -> PResult<Vec<u8>, &[u8]> {
    bytes::u8.array::<5>().ok_into().parse(input)
}

assert!(my_parser.parse(b"hello world") == Ok(Success(Vec::from(b"hello"), b" world")));

fn err_into<E>(self) -> impl Parse<I, Parsed = Self::Parsed, Error = E>
where Self: Sized, Self::Error: Into<E>, E: Error<I>,

Creates a parser whose parsing error is converted via Into.

If slef fails to parse, the returned parsing error is converted to E using the Into<E> trait.

See also basic::err_into.

Example

#[derive(Debug, PartialEq)]
struct MyError<'a>(&'a [u8]);

impl<'a> Error<&'a [u8]> for MyError<'a> { ... }
impl<'a> From<bytes::Error<&'a [u8]>> for MyError<'a> { ... }

fn my_parser(input: &[u8]) -> PResult<u8, &[u8], MyError<'_>> {
    bytes::u8.err_into().parse(input)
}

assert!(my_parser.parse(b"hello") == Ok(Success(b'h', b"ello")));
assert!(my_parser.parse(b"") == Err(Failure(MyError(b""), b"")));

fn res_try_into<R, E>(self) -> impl Parse<I, Parsed = R, Error = E>
where Self: Sized, Self::Parsed: TryInto<R>, Self::Error: Into<E>, E: Error<I>,

Creates a parser whose parsed result is converted via TryInto.

If self parses successfully, its result value is converted to R using the TryInto<R> trait. If TryInto fails, a parsing error is returned by calling E::invalid_input. If self returns a parsing error, that error is converted to E using the Into<E> trait.

Note that the parsing error is not converted using TryInto, but rather the infallible Into trait. Only the successfully parsed result converts via TryInto.

See also basic::res_try_into.

Example

#[derive(Debug, PartialEq)]
struct MyError<'a>(&'a [u8]);

impl<'a> Error<&'a [u8]> for MyError<'a> { ... }
impl<'a> From<bytes::Error<&'a [u8]>> for MyError<'a> { ... }

fn my_parser(input: &[u8]) -> PResult<char, &[u8], MyError<'_>> {
    bytes::be::u32.res_try_into().parse(input)
}

assert!(my_parser.parse(b"\x00\x00\x00hello") == Ok(Success('h', b"ello")));
assert!(my_parser.parse(b"\xff\xff\xffhello")
    == Err(Failure(MyError(b"\xff\xff\xffhello"), b"\xff\xff\xffhello")));
assert!(my_parser.parse(b"") == Err(Failure(MyError(b""), b"")));

fn ok_try_into<R>(self) -> impl Parse<I, Parsed = R, Error = Self::Error>
where Self: Sized, Self::Parsed: TryInto<R>,

Creates a parser whose parsed result is converted via TryInto.

If self parses successfully, its result value is converted to R using the TryInto<R> trait. If TryInto fails, a parsing error is returned by calling Error::invalid_input.

See also basic::ok_try_into.

Example

fn my_parser(input: &[u8]) -> PResult<char, &[u8]> {
    bytes::be::u32.ok_try_into().parse(input)
}

assert!(my_parser.parse(b"\x00\x00\x00hello") == Ok(Success('h', b"ello")));
assert!(my_parser.parse(b"\xff\xff\xffhello").is_err());

fn with<F, T>( self, with_fn: F, ) -> impl Parse<I, Parsed = T, Error = Self::Error>
where Self: Sized, F: Fn() -> T,

Converts a parser to produce a different return value.

If self parses successfully, its returned value is immediately dropped. with_fn is then called to produce the new value. If self fails, with_fn is not called.

See also basic::with.

Example

use pars::unicode::strict::verbatim;

#[derive(Debug, PartialEq)]
enum Token {
    Begin,
    End,
}

fn begin(input: &str) -> PResult<Token, &str> {
    verbatim("BEGIN").with(|| Token::Begin).parse(input)
}

fn end(input: &str) -> PResult<Token, &str> {
    verbatim("END").with(|| Token::End).parse(input)
}

assert_eq!(begin.parse("BEGIN hello"), Ok(Success(Token::Begin, " hello")));
assert_eq!(end.parse("END world"), Ok(Success(Token::End, " world")));

fn with_default<T>(self) -> impl Parse<I, Parsed = T, Error = Self::Error>
where Self: Sized, T: Default,

Converts a parser to produce a different return value, via Default.

If self parses successfully, its returned value is immediately dropped. The old returned value is replaced with a value of type T, produced by calling Default::default().

parser.with_default() is equivalent to parser.with(Default::default) (see Parse::with).

See also Parse::with_default.

Example

use pars::unicode::strict::verbatim;

#[derive(Debug, PartialEq, Default)]
struct Token;

fn token(input: &str) -> PResult<Token, &str> {
    verbatim("TOKEN").with_default().parse(input)
}

assert_eq!(token.parse("TOKEN"), Ok(Success(Token, "")));

fn with_value<T>( self, value: T, ) -> impl Parse<I, Parsed = T, Error = Self::Error>
where Self: Sized, T: Clone,

Converts a parser to produce a different return value, via Clone.

If self parses successfully, its returned value is immediately dropped. value is is cloned to replace the returned value.

See also basic::with_value.

Example

use pars::unicode::strict::verbatim;

#[derive(Debug, PartialEq, Clone)]
struct Token;

fn token(input: &str) -> PResult<Token, &str> {
    verbatim("TOKEN").with_value(Token).parse(input)
}

assert_eq!(token.parse("TOKEN"), Ok(Success(Token, "")));

fn flat_map<F, P>( self, combinator: F, ) -> impl Parse<I, Parsed = P::Parsed, Error = Self::Error>
where Self: Sized, F: Fn(Self::Parsed) -> P, P: Parse<I, Error = Self::Error>,

Maps the result of a parser onto a combinator to produce a new parser.

If self parses successfully, the result value is passed into combinator. The remaining input after self succeeds is then parsed by the parser returned by combinator.

Generally, this combinator is useful for building a parser that should parse differently depending on information earlier in the input stream.

See also basic::flat_map.

Example

// First byte is the length of the string, then that many
// more bytes is the string data.
fn pascal_str(input: &[u8]) -> PResult<&[u8], &[u8]> {
    bytes::u8.ok_into().flat_map(take).ok_into().parse(input)
}

assert!(pascal_str.parse(b"\x05hello") == Ok(Success(b"hello", b"")));
assert!(pascal_str.parse(b"\x05hi").is_err());

fn try_flat_map<F, P, S>( self, combinator: F, ) -> impl Parse<I, Parsed = P::Parsed, Error = Self::Error>
where Self: Sized, S: ErrorSeed<I, Self::Error>, F: Fn(Self::Parsed) -> Result<P, S>, P: Parse<I, Error = Self::Error>,

Maps the result of a parser onto a combinator to produce a new parser.

If self parses successfully, the result value is passed into combinator. combinator then returns a Result containing either a new parser or an error implementing ErrorSeed. If a parser is returned, the remaining input from self is passed to the new parser. Otherwise, the error is converted to a parsing error and returned.

See also basic::try_flat_map.

Example

fn my_parser(input: &[u8]) -> PResult<&[u8], &[u8]> {
    bytes::u8.try_flat_map(|value| {
        if value < 10 {
            Ok(take(value.into()).ok_into())
        } else {
            Err(ErrorKind::InvalidInput)
        }
    }).parse(input)
}

assert!(my_parser.parse(b"\x05hello") == Ok(Success(b"hello", b"")));
assert!(my_parser.parse(b"\x0bhello world").is_err());

fn or_value( self, value: Self::Parsed, ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
where Self: Sized, Self::Parsed: Clone,

Creates a parser that returns a fallback value instead of an error.

If self fails to parse, value is cloned and returned as a success value. The new parser will never return a parsing error.

See also basic::or_value

Example

fn my_parser(input: &str) -> PResult<char, &str> {
    unicode::strict::char.or_value('\0').parse(input)
}

assert_eq!(my_parser.parse("hello"), Ok(Success('h', "ello")));
assert_eq!(my_parser.parse(""), Ok(Success('\0', "")));

fn or_default(self) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
where Self: Sized, Self::Parsed: Default,

Creates a parser that returns a fallback value instead of an error.

If self fails to parse, a default value is returned by calling Default::default instead of returning the error. The new parser will never return a parsing error.

See also basic::or_default.

Example

fn my_parser(input: &str) -> PResult<char, &str> {
    unicode::strict::char.or_default().parse(input)
}

assert_eq!(my_parser.parse("hello"), Ok(Success('h', "ello")));
assert_eq!(my_parser.parse(""), Ok(Success('\0', "")));

fn or_else<F>( self, or_else_fn: F, ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
where Self: Sized, F: Fn() -> Self::Parsed,

Creates a parser that returns a fallback value instead of an error.

If self fails to parse, or_else_fn is called and its return value is returned as a success value. The new parser will never return a parsing error.

See also basic::or_else

Example

fn my_parser(input: &str) -> PResult<char, &str> {
    unicode::strict::char.or_else(|| '\0').parse(input)
}

assert_eq!(my_parser.parse("hello"), Ok(Success('h', "ello")));
assert_eq!(my_parser.parse(""), Ok(Success('\0', "")));

fn repeated_until<P, F, R>( self, sentinel: P, collect_fn: F, ) -> impl Parse<I, Parsed = (R, P::Parsed), Error = Self::Error>
where Self: Sized, P: Parse<I, Error = Self::Error>, F: for<'a> Fn(RepeatedUntilIter<'a, Self, P, I>) -> R,

Creates a parser that is repeated until a sentinel is parsed.

Parse::repeated_until produces a parser that will apply self repeatedly until sentinel_parser succeeds. That is, sentinel_parser is applied; if it fails, self is applied before trying sentinel_parser again. If sentinel_parser succeeds, parsing is complete.

The values produced by self are passed to collect_fn in the form of an iterator, RepeatedUntilIter. The value returned by collect_fn and the parsed value from sentinel_parser together as a 2-tuple become the parsed result of the overall new parser. If parser returns a parsing error at any point, a parsing error is returned from the overall new parser.

Note that the returned new parser does not allocate. Values produced by the iterator are obtained on demand by applying self each time Iterator::next is called. Allocation will only occur if the user provided function collect_fn allocates to produce its result.

See also basic::repeated_until.

Example

fn c_str_len(input: &[u8]) -> PResult<usize, &[u8]> {
    bytes::u8.repeated_until(bytes::verbatim(b"\x00"), |iter| iter.count())
        .map(|(len, _)| len) // we only want the length
        .parse(input)
}

assert!(c_str_len.parse(b"hello\x00world") == Ok(Success(5, b"world")));
assert!(c_str_len.parse(b"\x00") == Ok(Success(0, b"")));
assert!(c_str_len.parse(b"hello").is_err());

fn try_repeated_until<P, F, R, S>( self, sentinel: P, collect_fn: F, ) -> impl Parse<I, Parsed = (R, P::Parsed), Error = Self::Error>
where Self: Sized, S: ErrorSeed<I, Self::Error>, P: Parse<I, Error = Self::Error>, F: for<'a> Fn(RepeatedUntilIter<'a, Self, P, I>) -> Result<R, S>,

Creates a parser that is repeated until a sentinel is parsed.

Parse::try_repeated_until produces a parser that will apply self repeatedly until sentinel_parser succeeds. That is, sentinel_parser is applied; if it fails, self is applied before trying sentinel_parser again. If sentinel_parser succeeds, parsing is complete.

The values produced by self are passed to collect_fn in the form of an iterator, RepeatedUntilIter. If collect_fn returns an Ok value, that and the value from sentinel_parser together as a 2-tuple become the parsed result of the overall new parser. If collect_fn returns an Err, a parsing error is returned. If self returns a parsing error at any point, a parsing error is returned form the overall new parser.

Note that the returned new parser does not allocate. Values produced by the iterator are obtained on demand by applying self each time Iterator::next is called. Allocation will only occur if the user provided function collect_fn allocates to produce its result.

See also basic::try_repeated_until.

Example

fn c_str_len(input: &[u8]) -> PResult<usize, &[u8]> {
    bytes::u8.try_repeated_until(bytes::verbatim(b"\x00"), |iter| {
        let count = iter.count();
        // disallow empty strings
        if count == 0 {
            Err(ErrorKind::InvalidInput)
        } else {
            Ok(count)
        }
    }).map(|(len, _)| len) // we only want the length
        .parse(input)
}

assert!(c_str_len.parse(b"hello\x00world") == Ok(Success(5, b"world")));
assert!(c_str_len.parse(b"\x00").is_err());
assert!(c_str_len.parse(b"hello").is_err());

fn collect_repeated_until<P, R>( self, sentinel: P, ) -> impl Parse<I, Parsed = (R, P::Parsed), Error = Self::Error>
where Self: Sized, P: Parse<I, Error = Self::Error>, R: FromIterator<Self::Parsed>,

Creates a parser that is repeated until a sentinel is parsed.

Parse::collect_repeated_until produces a parser that will apply self repeatedly until sentinel_parser succeeds. That is, sentinel_parser is applied; if it fails, self is applied before trying sentinel_parser again. If sentinel_parser succeeds, parsing is complete.

The values produced by self are collected using the FromIterator trait implementation for the new parser’s parsed result type, which is usually deduced. This value together with the parsed value from sentinel_parser as a 2-tuple become the parsed value for the overall new parser. If self returns a parsing error at any point, a parsing error is returned from the overall new parser.

Note that the returned new parser does not allocate unless the FromIterator implementation allocates.

See also basic::collect_repeated_until.

Example

fn c_str(input: &[u8]) -> PResult<Vec<u8>, &[u8]> {
    bytes::u8.collect_repeated_until(bytes::verbatim(b"\x00"))
        .map(|(s, _)| s)
        .parse(input)
}

assert!(c_str.parse(b"hello\x00world") == Ok(Success(Vec::from(b"hello"), b"world")));
assert!(c_str.parse(b"\x00") == Ok(Success(Vec::new(), b"")));
assert!(c_str.parse(b"hello").is_err());

fn repeated<N, F, R>( self, range: N, collect_fn: F, ) -> impl Parse<I, Parsed = R, Error = Self::Error>
where Self: Sized, N: Range, F: for<'a> Fn(RepeatedIter<'a, Self, I>) -> R,

Creates a parser that is repeated a number of times within a range.

Parse::repeated produces a parser that will apply self repeatedly. The number of times self will be applied is defined by range. For example, if range is 3..=5, self will be applied at least 3 times, and will be applied up to 5 times if possible. If self fails before parsing 3 times (or whatever lower bound applies to range, if any), a parsing error is returned from the new parser. Once the upper bound is met (again, if any) parsing automatically ends successfully. self returning a parsing error after meeting the lower bound also ends parsing for the new parser.

The values produced by self are passed to collect_fn in the form of an iterator, RepeatedIter. Whever collect_fn returns is the parsed value of the new parser.

Note that the returned new parser does not allocate. Values produced by the iterator are obtained on demand by applying self each time Iterator::next is called. Allocation will only occur if the user provided function collect_fn allocates to produce its result.

See also basic::repeated.

Example

fn my_parser(input: &[u8]) -> PResult<u32, &[u8]> {
    bytes::u8.repeated(2..=4, |iter| {
        iter.map(|byte| byte as u32).sum()
    }).parse(input)
}

assert!(my_parser.parse(b"\x01\x02\x03\x04\x05") == Ok(Success(10, b"\x05")));
assert!(my_parser.parse(b"\x01\x02\x03") == Ok(Success(6, b"")));
assert!(my_parser.parse(b"\x01\x02") == Ok(Success(3, b"")));
assert!(my_parser.parse(b"\x01").is_err());
assert!(my_parser.parse(b"").is_err());

fn try_repeated<N, F, R, S>( self, range: N, collect_fn: F, ) -> impl Parse<I, Parsed = R, Error = Self::Error>
where Self: Sized, N: Range, S: ErrorSeed<I, Self::Error>, F: for<'a> Fn(RepeatedIter<'a, Self, I>) -> Result<R, S>,

Creates a parser that is repeated a number of times within a range.

Parse::try_repeated produces a parser that will apply self repeatedly. The number of times self will be applied is defined by range. For example, if range is 3..=5, self will be applied at least 3 times, and will be applied up to 5 times if possible. If self fails before parsing 3 times (or whatever lower bound applies to range, if any), a parsing error is returned from the new parser. Once the upper bound is met (again, if any) parsing automatically ends successfully. self returning a parsing error after meeting the lower bound also ends parsing for the new parser.

The values produced by self are passed to collect_fn in the form of an iterator, RepeatedIter. If collect_fn returns an Ok value, the contained value becomes the parsed value of the new parser. If collect_fn returns an Err, that gets returned as a parsing error via ErrorSeed.

Note that the returned new parser does not allocate. Values produced by the iterator are obtained on demand by applying self each time Iterator::next is called. Allocation will only occur if the user provided function collect_fn allocates to produce its result.

See also basic::try_repeated.

Example

fn my_parser(input: &[u8]) -> PResult<u32, &[u8]> {
    bytes::u8.try_repeated(2..=4, |iter| {
        let sum = iter.map(|byte| byte as u32).sum();
        // only allow even sums
        if sum % 2 == 0 {
            Ok(sum)
        } else {
            Err(ErrorKind::InvalidInput)
        }
    }).parse(input)
}

assert!(my_parser.parse(b"\x01\x02\x03\x04\x05") == Ok(Success(10, b"\x05")));
assert!(my_parser.parse(b"\x01\x02\x03") == Ok(Success(6, b"")));
assert!(my_parser.parse(b"\x01\x02").is_err());
assert!(my_parser.parse(b"\x01").is_err());
assert!(my_parser.parse(b"").is_err());

fn collect_repeated<C>( self, range: impl Range, ) -> impl Parse<I, Parsed = C, Error = Self::Error>
where Self: Sized, C: FromIterator<Self::Parsed>,

Creates a parser that is repeated a number of times within a range.

Parse::collect_repeated produces a parser that will apply self repeatedly. The number of times self will be applied is defined by range. For example, if range is 3..=5, self will be applied at least 3 times, and will be applied up to 5 times if possible. If self fails before parsing 3 times (or whatever lower bound applies to range, if any), a parsing error is returned from the new parser. Once the upper bound is met (again, if any) parsing automatically ends successfully. self returning a parsing error after meeting the lower bound also ends parsing for the new parser.

The values produced by self are collected by the FromIterator trait implementation on the parsed type of the parser, which is usually deduced.

Note that the returned new parser does not allocate unless the FromIterator implementation allocates.

See also basic::collect_repeated.

Example

use pars::unicode::{strict::char_with_prop, prop::Alphabetic};

fn my_parser(input: &str) -> PResult<String, &str> {
    char_with_prop(Alphabetic).collect_repeated(3..=5)
        .parse(input)
}

assert_eq!(my_parser.parse("hello world"), Ok(Success(String::from("hello"), " world")));
assert_eq!(my_parser.parse("hey there"), Ok(Success(String::from("hey"), " there")));
assert!(my_parser.parse("hi there").is_err());
assert_eq!(my_parser.parse("goodbye"), Ok(Success(String::from("goodb"), "ye")));

fn array<const LEN: usize>( self, ) -> impl Parse<I, Parsed = [Self::Parsed; LEN], Error = Self::Error>
where Self: Sized,

Creates a parser that is repeated a constant number of times.

Parse::array produces a parser that applies self LEN times and places the parsed results in an array. If self returns a parsing error before parsing LEN times, a parsing error is returned from the new parser.

Parse::array is similar to Parse::repeated, except that the number of repetitions is known at compile time and the results are placed in array rather than being processed via an iterator. Most cases where the number of repetitions is known at compile time should prefer to use Parse::array, but be aware of the tradeoffs of storing and moving the values on the stack when LEN is large.

Example

fn my_parser(input: &[u8]) -> PResult<[u8; 4], &[u8]> {
    bytes::u8.array().parse(input)
}

assert!(my_parser.parse(b"\x01\x02\x03\x04\x05") == Ok(Success([1, 2, 3, 4], b"\x05")));
assert!(my_parser.parse(b"\x01\x02\x03").is_err());

fn separated<N, S, F, R>( self, range: N, separator: S, collect_fn: F, ) -> impl Parse<I, Parsed = R, Error = Self::Error>
where Self: Sized, N: Range, S: Parse<I, Error = Self::Error>, F: for<'a> Fn(SeparatedIter<'a, Self, S, I>) -> R,

Creates a parser that is applied repeatedly while interspersed with a separator.

Parse::separated produces a parser that will apply self and separator repeatedly, alternatiging between the two. The parsing pattern starts with parser and ends with self, such as self separator self. This is distinct from parser.then(separator).repeated(...) as Parse::separated will not parse a trailing separator, and a repeated pair must parse a trailing separator.

The provided range determines the minimum and maximum number of repatitions of self. For example if range is 3..=5, self must be able to parse at least 3 times, and up to 5 times. Similarly, .. corresponds to 0 or more repetitions, and 1.. corresponds to 1 or more. An exact number of repetitions can be specified with a single integer value such as 3 (but 3..=3 is also valid and equivalent). See Range for more information on ranges.

The values produced by self are passed to collect_fn in the form of an iterator, SeparatedIter. Whatever collect_fn returns is the parsed value of the new parser. The values produced by separator are discarded.

Note that the returned new parser does not allocate. Values produced by the iterator are obtained on demand by applying separator and self each time Iterator::next is called. Allocation will only occur if the user provided collect_fn allocates when called.

See also basic::separated.

Example

fn my_parser(input: &str) -> PResult<Vec<char>, &str> {
    unicode::strict::char.separated(
        ..,
        unicode::strict::verbatim(","),
        |iter| {
            let mut out = Vec::new();
            for ch in iter {
                out.push(ch);
            }
            out
        }
    ).parse(input)
}

assert_eq!(
    my_parser.parse("a,b,c"),
    Ok(Success(vec!['a', 'b', 'c'], "")),
);

assert_eq!(
    my_parser.parse("a,b,c,"),
    Ok(Success(vec!['a', 'b', 'c'], ",")),
);

fn try_separated<N, S, F, R, E>( self, range: N, separator: S, collect_fn: F, ) -> impl Parse<I, Parsed = R, Error = Self::Error>
where Self: Sized, N: Range, S: Parse<I, Error = Self::Error>, F: for<'a> Fn(SeparatedIter<'a, Self, S, I>) -> Result<R, E>, E: ErrorSeed<I, Self::Error>,

Creates a parser that is applied repeatedly while interspersed with a separator.

Parse::try_separated produces a parser that will apply self and separator repeatedly, alternatiging between the two. The parsing pattern starts with self and ends with self, such as self separator self. This is distinct from parser.then(separator).try_repeated(...) as Parse::try_separated will not parse a trailing separator, and a repeated pair must parse a trailing separator.

The provided range determines the minimum and maximum number of repatitions of self. For example if range is 3..=5, self must be able to parse at least 3 times, and up to 5 times. Similarly, .. corresponds to 0 or more repetitions, and 1.. corresponds to 1 or more. An exact number of repetitions can be specified with a single integer value such as 3 (but 3..=3 is also valid and equivalent). See Range for more information on ranges.

The values produced by self are passed to collect_fn in the form of an iterator, SeparatedIter. collect_fn then returns a Result, where the Ok variant becomes the parsed value of the new parser, and the Err variant becomes an error returned by the new parser via the ErrorSeed trait.

Note that the returned new parser does not allocate. Values produced by the iterator are obtained on demand by applying separator and self each time Iterator::next is called. Allocation will only occur if the user provided collect_fn allocates when called.

See also basic::try_separated.

Example

fn my_parser(input: &str) -> PResult<Vec<char>, &str> {
    unicode::strict::char.try_separated(
        ..,
        unicode::strict::verbatim(","),
        |iter| -> Result<Vec<char>, ErrorKind> {
            let mut out = Vec::new();
            for ch in iter {
                if !ch.is_ascii_alphabetic() {
                    return Err(ErrorKind::InvalidInput);
                }
                out.push(ch);
            }
            Ok(out)
        }
    ).parse(input)
}

assert_eq!(
    my_parser.parse("a,b,c"),
    Ok(Success(vec!['a', 'b', 'c'], "")),
);

assert_eq!(
    my_parser.parse("a,b,c,"),
    Ok(Success(vec!['a', 'b', 'c'], ",")),
);

assert!(my_parser.parse("a,2,c").is_err());

fn collect_separated<C>( self, range: impl Range, separator: impl Parse<I, Error = Self::Error>, ) -> impl Parse<I, Parsed = C, Error = Self::Error>
where Self: Sized, C: FromIterator<Self::Parsed>,

Creates a parser that is applied repeatedly while interspersed with a separator.

Parse::collect_separated produces a parser that will apply self and separator repeatedly, alternatiging between the two. The parsing pattern starts with self and ends with self, such as self separator self. This is distinct from parser.then(separator).collect_repeated(...) as Parse::collect_separated will not parse a trailing separator, and a repeated pair must parse a trailing separator.

The provided range determines the minimum and maximum number of repatitions of self. For example if range is 3..=5, self must be able to parse at least 3 times, and up to 5 times. Similarly, .. corresponds to 0 or more repetitions, and 1.. corresponds to 1 or more. An exact number of repetitions can be specified with a single integer value such as 3 (but 3..=3 is also valid and equivalent). See Range for more information on ranges.

The values produced by self are collected via the FromIterator trait on the, typically inferred, parsed value type of the new parser. The values produced by separator are discarded.

See also basic::collect_separated.

Example

fn my_parser(input: &str) -> PResult<Vec<char>, &str> {
    // Note that `Vec<char>` is inferred from the return type of `my_parser`
    unicode::strict::char.collect_separated(
        ..,
        unicode::strict::verbatim(","),
    ).parse(input)
}

assert_eq!(
    my_parser.parse("a,b,c"),
    Ok(Success(vec!['a', 'b', 'c'], "")),
);

assert_eq!(
    my_parser.parse("a,b,c,"),
    Ok(Success(vec!['a', 'b', 'c'], ",")),
);

fn complete(self) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
where Self: Sized,

Creates a parser that fails if all input is not consumed.

Parse::complete produces a parser that applies self and then verifies that all input was consumed. If there is still input remaining, an error is returned via Error::expected_eof.

See also basic::complete.

Example

fn my_parser(input: &[u8]) -> PResult<u32, &[u8]> {
    bytes::be::u32.complete().parse(input)
}

assert!(my_parser.parse(b"\x01\x02\x03\x04") == Ok(Success(0x01020304, &[][..])));
assert!(my_parser.parse(b"\x01\x02\x03\x04\x05").is_err());

fn spanned( self, ) -> impl Parse<I, Parsed = (Self::Parsed, Span<I>), Error = Self::Error>
where Self: Sized,

Creates a parser that returns a Span of parsed input in addition to the parsed value.

Parse::spanned produces a parser that applies self, and on success, it returns the parsed value from parser in a tuple with a Span of the input consumed to produce the parsed value.

If only the Span is needed and the parsed value can be discarded, Parse::recognize can also be used instead.

See also basic::spanned.

Example

fn my_parser(input: &[u8]) -> PResult<(u32, Span<&[u8]>), &[u8]> {
    bytes::be::u32.spanned().parse(input)
}

let res = my_parser.parse(b"\x01\x02\x03\x04\x05");
assert!(res.is_ok());
let Success((value, span), rem) = res.unwrap();
let span: &[u8] = span.into();
assert_eq!(value, 0x01020304);
assert_eq!(span, &[1u8, 2, 3, 4][..]);
assert_eq!(rem, &[5u8][..]);

fn spanned_into<S>( self, ) -> impl Parse<I, Parsed = (Self::Parsed, S), Error = Self::Error>
where Self: Sized, Span<I>: Into<S>,

Creates a parser that also returns a span of the input that was parsed.

Parse::spanned_into produces a parser that combines the parsed result of self with a Span of the input that self parsed converted to S via Into in a 2-tuple. This allows a parser to provide information about where in an input stream a parsed value originated from.

Parse::spanned_into differs from Parse::spanned in that the produced Span is converted into another type, S by way of Into. This is a common enough pattern that Parse::spanned_into exists to simplify usage. For example, a Span<&str> can be converted to a &str that contains just the content of the span.

If the parsed value returned by self is not needed, such as a parser that returns the unit type, consider using Parse::recognize instead. Note that Parse::recognize doesn’t provide a recognize_into variant, since the same effect can be accomplished with parser.recognize().ok_into() in this case.

See also basic::spanned_into.

Example

fn my_parser(input: &[u8]) -> PResult<(u32, &[u8]), &[u8]> {
    bytes::be::u32.spanned_into().parse(input)
}

assert!(my_parser.parse(b"\x01\x02\x03\x04\x05") ==
    Ok(Success((0x01020304, b"\x01\x02\x03\x04"), b"\x05")));
assert!(my_parser.parse(b"").is_err());

fn recognize(self) -> impl Parse<I, Parsed = Span<I>, Error = Self::Error>
where Self: Sized,

Creates a parser that returns a Span of parsed input.

Parse::recognize produces a parser that applies self and on success, it returns a Span of the input consumed by self.

Parse::recognize is similar to Parse::spanned, except that the parsed value returned by self is discarded.

See also basic::recognize.

Example

fn my_parser(input: &[u8]) -> PResult<Span<&[u8]>, &[u8]> {
    bytes::be::u32.recognize().parse(input)
}

let res = my_parser.parse(b"\x01\x02\x03\x04\x05");
assert!(res.is_ok());
let Success(span, rem) = res.unwrap();
let span: &[u8] = span.into();
assert_eq!(span, b"\x01\x02\x03\x04");
assert_eq!(rem, b"\x05");

fn cond( self, condition: bool, ) -> impl Parse<I, Parsed = Option<Self::Parsed>, Error = Self::Error>
where Self: Sized,

Creates a parser that produces a parsed value only if condition is true.

Parse::cond produces a parser that applies self only if condition is true. On success, the parsed value returned by self returned wrapped in Some. If condition is false, the parser succeeds, but returns None as the parsed value.

See also basic::cond.

Example

fn my_parser(input: &[u8]) -> PResult<Option<u32>, &[u8]> {
    bytes::u8.flat_map(|b| {
        bytes::be::u32.cond(b != 0)
    }).parse(input)
}

assert!(my_parser.parse(b"\x00\x01\x02\x03\x04") == Ok(Success(None, b"\x01\x02\x03\x04")));
assert!(my_parser.parse(b"\x01\x01\x02\x03\x04") == Ok(Success(Some(0x01020304), b"")));

fn verify<F>( self, verify_fn: F, ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
where Self: Sized, F: Fn(&Self::Parsed) -> bool,

Creates a parser that succeeds only if the parsed value passes the given predicate.

Parse::verify produces a parser that applies self, and on success, passes the parsed value to verify_fn. If verify_fn returns true, the parser returns the parsed value. If verify_fn returns false, the parser returns an error via Error::invalid_input.

See also basic::verify.

Example

fn non_zero(input: &[u8]) -> PResult<u32, &[u8]> {
    bytes::be::u32.verify(|v| *v != 0).parse(input)
}

assert!(non_zero.parse(b"\x00\x00\x00\x00").is_err());
assert!(non_zero.parse(b"\x00\x00\x00\x01") == Ok(Success(1, b"")));

fn not(self) -> impl Parse<I, Parsed = (), Error = Self::Error>
where Self: Sized,

Creates a parser that succeeds only if self fails.

Parse::not produces a parser that applies self, an on success, it returns an error via Error::invalid_input. If self fails, the new parser returns a success.

See also basic::not.

Example

fn my_parser(input: &str) -> PResult<(), &str> {
    ascii::strict::char_with_prop(ascii::prop::Printable).not().parse(input)
}

assert!(my_parser.parse("hello").is_err());
assert!(my_parser.parse("\0").is_ok());

fn opt( self, ) -> impl Parse<I, Parsed = Option<Self::Parsed>, Error = Self::Error>
where Self: Sized,

Creates a parser that attempts to parse with self if possible.

Parse::opt produces a parser that applies self, and on success, the new parser returns the parsed value wrapped in Some. On failure, the new parser still succeeds but returns None without consuming any input.

See also basic::opt.

Example

fn my_parser(input: &[u8]) -> PResult<Option<&[u8]>, &[u8]> {
    bytes::verbatim("hello").ok_into().opt().parse(input)
}

assert!(my_parser.parse(b"hello world") == Ok(Success(Some(&b"hello"[..]), b" world")));
assert!(my_parser.parse(b"goodbye world") == Ok(Success(None, b"goodbye world")));

fn peek(self) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
where Self: Sized,

Creates a parser that parses without consuming input.

Parse::peek produces a parser that applies self, and on success, the new parser returns the parsed value returned by self and the entire input that was passed to the parser. On failure, the new parser forwards the error unmodified.

See also basic::peek.

Example

fn my_parser(input: &[u8]) -> PResult<u32, &[u8]> {
    bytes::be::u32.peek().parse(input)
}

assert!(my_parser.parse(b"\x01\x02\x03\x04") == Ok(Success(0x01020304, b"\x01\x02\x03\x04")));

fn by_ref<'a>( &'a self, ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error> + 'a
where I: 'a,

Creates a parser out of a reference to another parser.

Parse::by_ref produces a parser that directly applies self. The produced parser behaves identically to self and internally contains a reference to self.

Parse is not automatically implemented for references to types that implement Parse, so Parse::by_ref can by used to convert a reference to a parser into a parser.

See also basic::by_ref.

Example

fn my_parser(input: &[u8]) -> PResult<u8, &[u8]> {
    bytes::u8.by_ref().parse(input)
}

assert!(my_parser.parse(b"\x01\x02\x03\x04") == Ok(Success(1u8, b"\x02\x03\x04")));
assert!(my_parser.parse(b"").is_err());

fn then<P>( self, other: P, ) -> impl Parse<I, Parsed = (Self::Parsed, P::Parsed), Error = Self::Error>
where Self: Sized, P: Parse<I, Error = Self::Error>,

Creates a parser that applies two parsers in sequence.

Parse::then produces a parser that applies self and other in that order. On success, the new parser returns the parsed values returned by self and other as a tuple. If either parser fails, the corresponding error is returned from the new parser. When self fails, other is not applied.

See also basic::pair and basic::seq.

Example

fn my_parser(input: &str) -> PResult<(char, char), &str> {
    unicode::strict::char.then(unicode::strict::char)
        .parse(input)
}

assert_eq!(my_parser.parse("abc"), Ok(Success(('a', 'b'), "c")));
assert!(my_parser.parse("a").is_err());
assert!(my_parser.parse("").is_err());

fn or<P>( self, other: P, ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
where Self: Sized, P: Parse<I, Parsed = Self::Parsed, Error = Self::Error>,

Creates a parser that attempts to parse with each of two parsers.

Parse::or produces a parser that first attempts to apply self and then applies other if self fails. If self succeeds, the parsed value self returns is returned by the new parser. If self fails and then other succeeds, the parsed value returned by other is returned by the new parser. If both fail, the error returned by other is returned.

See also basic::either and basic::alt.

Example

fn my_parser(input: &str) -> PResult<&str, &str> {
    either(
        unicode::strict::verbatim("foo").ok_into(),
        unicode::strict::verbatim("bar").ok_into(),
    ).parse(input)
}

assert_eq!(my_parser.parse("foobaz"), Ok(Success("foo", "baz")));
assert_eq!(my_parser.parse("barbaz"), Ok(Success("bar", "baz")));
assert!(my_parser.parse("baz").is_err());

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.