pars
Type Alias PResult

pub type PResult<T, I, E> = Result<Success<T, I>, Failure<E, I>>;

The Result type returned by a parser.

All parsers return either a Success when parsing succeeds, or a Failure when parsing fails. The input type that is returned is the same in either case, and is the same type as was provided to the parser to parse.

Unlike a typical Result alias in rust, PResult has three generic parameters: The parsed value type, the input type, and the error type. Since this differs from the standard convention, the alias has a slightly different name to clearly indicate the break from convention.

Aliased Type

enum PResult<T, I, E> {
    Ok(Success<T, I>),
    Err(Failure<E, I>),
}

Variants

Ok(Success<T, I>)

Contains the success value

Err(Failure<E, I>)

Contains the error value