pars

Module unicode

Available on crate feature unicode only.

Modules

lossy

Parsers that replace invalid Unicode input with the Unicode Replacement Character.

prop

Unicode property definitions and combinators.

strict

Parsers that treat any invalid Unicode as a parse error.

Structs

Error

LocInput

Unicode input that tracks location information.

Enums

ErrorKind

Traits

Property

UnicodeInput

UnicodeSymbol

Type Aliases

PResult