pars::unicode

Module lossy

Available on crate feature unicode only.

Parsers that replace invalid Unicode input with the Unicode Replacement Character.

Input that is not valid Unicode, such as invalid UTF-8 encodings or invalid code points, gets replaced by U+FFFD in the character stream. Thus these parsers can be lossy since some of the original information is lost in such cases. The benefit is that the input can potentially be parsed despite encoding errors or otherwise invalid input.

Macros

regex

Structs

CharWithPropParser

Functions

char

char_with_prop

grapheme_cluster

line

verbatim