pars::ascii
Macro ascii

ascii!() { /* proc-macro */ }

Create an ASCII character or string, checked at compile time.

This macro takes a single literal argument and converts the value to an AsciiChar or an &'static AsciiStr depending on the literal type. The argument can be an integer, character, string, byte, or byte string literal. If the literal is not valid ASCII, a compile error is emitted.

This macro can be used in const contexts.

Example

assert_eq!(ascii!(0x41), AsciiChar::A);
assert_eq!(ascii!('A'), AsciiChar::A);
assert_eq!(ascii!(b'A'), AsciiChar::A);
assert_eq!(ascii!("ABC").as_slice(), &[AsciiChar::A, AsciiChar::B, AsciiChar::C]);
assert_eq!(ascii!(b"ABC").as_slice(), &[AsciiChar::A, AsciiChar::B, AsciiChar::C]);