pars
Module basic

Generic parser and combinator building blocks.

This module provides the majority of the combinators that will be used to implement most parsers. Everything implemented in this module supports any combination of input and error types. When using pars, it is best to become familiar with the contents of this module, as the majority of parsing problems should have a solution implemented here.

Most functions that take a parser as a parameter and produce a transformed parser (A.K.A. combinators) are also available as methods on the Parse trait.

Macros

alt

Creates a parser that will match exactly one of its arguments.

permutation

seq

Creates a parser that matches a sequence of parsers.

Structs

Many0Iter

An Iterator over repeated applications of a parser.

Many1Iter

An Iterator over repeated applications of a parser.

ManyIter

An Iterator over repeated applications of a parser.

ManyUntilIter

An Iterator over repeated applications of a parser.

RepeatedIter

An Iterator over repeated applications of a parser.

Functions

array

Creates a parser that is repeated a constant number of times.

as_ref

collect_many

Creates a parser that is repeated a number of times within a range.

collect_many0

Creates a parser that is repeated as many times as possible.

collect_many1

Creates a parser that is repeated at east once.

collect_many_until

Creates a parser that is repeated until a sentinel is parsed.

collect_repeated

Creates a parser that is repeated a set number of times.

complete

Creates a parser that must consume all input.

cond

constant

Creates a parser that unconditionally produces a specific value.

delimited

either

eof

A parser that succeeds on empty input, and fails otherwise.

err_into

Creates a parser whose parsing error is converted via Into.

error

Creates a parser that unconditionally produces an error.

flat_map

Maps the result of a parser onto a combinator to produce a new parser.

many

Creates a parser that is repeated a number of times within a range.

many0

Creates a parser that is repeated as many times as possible.

many1

Creates a parser that is repeated at least once.

many_until

Creates a parser that is repeated until a sentinel is parsed.

map

Creates a parser whose parsed result is transformed.

map_err

Creates a parser that transforms a parsing error.

map_res

Creates a parser whose result is transformed.

not

ok_into

Creates a parser whose parsed result is converted via Into.

ok_try_into

Creates a parser whose parsed result is converted via TryInto.

opt

or_default

Creates a parser that returns a fallback value instead of an error.

or_else

Creates a parser that returns a fallback value instead of an error.

or_value

Creates a parser that returns a fallback value instead of an error.

pair

peek

pop

A parser that extracts one symbol from the input stream.

prefix

recognize

remaining

remaining_len

repeated

Creates a parser that is repeated a set number of times.

res_into

Creates a parser whose parsed result or parsing error are converted via Into.

res_try_into

Creates a parser whose parsed result is converted via TryInto.

separated

spanned

Creates a parser that also returns a span of the input that was parsed.

spanned_into

Creates a parser that also returns a span of the input that was parsed.

suffix

take

Creates a parser that takes a number of symbols from the input.

try_flat_map

Maps the result of a parser onto a combinator to produce a new parser.

try_many

Creates a parser that is repeated a number of times within a range.

try_many0

Creates a parser that is repeated as many times as possible.

try_many1

Creates a parser that is repeated at least once.

try_many_until

Creates a parser that is repeated until a sentinel is parsed.

try_map

Creates a parser whose parsed result is fallibly transformed.

try_repeated

Creates a parser that is repeated a set number of times.

verbatim

Creates a parser that attempts to match input verbatim.

verify

with

Converts a parser to produce a different return value.

with_default

Converts a parser to produce a different return value, via Default.

with_value

Converts a parser to produce a different return value, via Clone.