pars::basic
Function array

pub const fn array<P, I, const LEN: usize>(
    parser: P,
) -> impl Parse<I, Parsed = [P::Parsed; LEN], Error = P::Error>
where
    P: Parse<I>,
    I: Input,

Creates a parser that is repeated a constant number of times.

array produces a parser that applies parser LEN times and places the parsed results in an array. If parser returns a parsing error before parsing LEN times, a parsing error is returned from the new parser.

array is similar to repeated, except that the number of repetitions is known at compile time and the results are placed in array rather than being processed via an iterator. Most cases where the number of repetitions is known at compile time should prefer to use array, but be aware of the tradeoffs of storing and moving the values on the stack when LEN is large.

See also Parse::array.

Example

fn my_parser(input: &[u8]) -> PResult<[u8; 4], &[u8]> {
    array(bytes::u8).parse(input)
}

assert!(my_parser.parse(b"\x01\x02\x03\x04\x05") == Ok(Success([1, 2, 3, 4], b"\x05")));
assert!(my_parser.parse(b"\x01\x02\x03").is_err());