pars::basic
Function res_try_into

pub const fn res_try_into<P, R, E, I>(
    parser: P,
) -> impl Parse<I, Parsed = R, Error = E>
where
    P: Parse<I>,
    I: Input,
    P::Parsed: TryInto<R>,
    P::Error: Into<E>,
    E: Error<I>,

Creates a parser whose parsed result is converted via TryInto.

If parser parses successfully, its result value is converted to R using the TryInto<R> trait. If TryInto fails, a parsing error is returned by calling E::invalid_input. If parser returns a parsing error, that error is converted to E using the Into<E> trait.

Note that the parsing error is not converted using TryInto, but rather the infallible Into trait. Only the successfully parsed result converts via TryInto.

See also Parse::res_try_into.

Example

#[derive(Debug, PartialEq)]
struct MyError<'a>(&'a [u8]);

impl<'a> Error<&'a [u8]> for MyError<'a> { ... }
impl<'a> From<bytes::Error<&'a [u8]>> for MyError<'a> { ... }

fn my_parser(input: &[u8]) -> PResult<char, &[u8], MyError<'_>> {
    res_try_into(bytes::be::u32).parse(input)
}

assert!(my_parser.parse(b"\x00\x00\x00hello") == Ok(Success('h', b"ello")));
assert!(my_parser.parse(b"\xff\xff\xffhello")
    == Err(Failure(MyError(b"\xff\xff\xffhello"), b"\xff\xff\xffhello")));
assert!(my_parser.parse(b"") == Err(Failure(MyError(b""), b"")));