pars
Module prelude

The pars prelude.

Re-exports

pub use super::basic::constant;

pub use super::basic::delimited;

pub use super::basic::either;

pub use super::basic::eof;

pub use super::basic::error;

pub use super::basic::pair;

pub use super::basic::pop;

pub use super::basic::prefix;

pub use super::basic::remaining;

pub use super::basic::remaining_len;

pub use super::basic::separated;

pub use super::basic::suffix;

pub use super::basic::take;

pub use super::Error as _;

pub use super::ErrorSeed as _;

pub use super::Failure;

pub use super::Input;

pub use super::IntoInput;

pub use super::PResultExt;

pub use super::Parse;

pub use super::Span;

pub use super::Success;

Macros

alt

Creates a parser that will match exactly one of its arguments.

permutation

seq

Creates a parser that matches a sequence of parsers.