pars
Trait Error

pub trait Error<I: Input>: Sized {
    // Required methods
    fn need_more_input(pos: I) -> Self;
    fn expected_eof(pos: I) -> Self;
    fn invalid_input(pos: I) -> Self;
    fn position(&self) -> &I;
}

A parsing error.

Required Methods

fn need_more_input(pos: I) -> Self

Creates an error representing the need for more input.

Generic parsers will call this method to generate an error when parsing failed due to reaching the end of input.

fn expected_eof(pos: I) -> Self

Creates an error representing that the end of input was expected but not reached.

Generic parsers will call this method to generate an error when parsing failed because it expected to reach the end of the input, but there was more input available after parsing was finished.

fn invalid_input(pos: I) -> Self

Creates an error representing that the input is unparsable.

Generic parsers will call this method to generate an error when parsing failed due to a failed verification of a symbol or parsed value.

fn position(&self) -> &I

Gets the input position where the error occured.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.