pars
Trait ErrorSeed

pub trait ErrorSeed<I: Input, E: Error<I>> {
    // Required method
    fn into_error(self, pos: I) -> E;
}

A parsing error descriptor that can be used to build a parsing error.

Most types implementing ErrorSeed are typically an “error kind” enum. The most generic [ErrorSeed] is [ErrorKind]. However, an [ErrorSeed] type can be anything as long as it can be combined with some [Input] to create an [Error`].

An ErrorSeed is generally used in fallible combinators (i.e. any Parse::try_* combinator) to represent an error with the input stream abstracted away. For example:

bytes::u8.try_map(|byte| {
    if byte < 0x80 {
        Ok(byte)
    } else {
        // `try_map` converts this to a `bytes::Error`
        Err(ErrorKind::InvalidInput)
    }
})

ErrorSeed is generic over both the input type and error type, so any ErrorSeed can convert multiple error types if desired. ErrorKind, for example, can convert to any type implementing Error for any input type.

Required Methods

fn into_error(self, pos: I) -> E

Combines this error seed with input to create a parsing error.