pars
Trait IntoInput

pub trait IntoInput {
    type Symbol;
    type Input: Input<Symbol = Self::Symbol>;

    // Required method
    fn into_input(self) -> Self::Input;
}

Trait for types that can be directly converted into an Input type.

IntoInput is analogous to IntoIterator. For example, &[u8] implements Input, but Vec<u8> does not. However, &Vec<u8> implements IntoInput and will convert to [&[u8]].

The Parse::parse function accepts a type implementing IntoInput for convenience. So given let my_input: Vec<u8>;, my_parser.parse(&my_input) works without needing to explicitly convert the Vec into a slice.

Required Associated Types

type Symbol

The symbol type of the Input this type will convert to.

type Input: Input<Symbol = Self::Symbol>

The Input this type will convert to.

Required Methods

fn into_input(self) -> Self::Input

Convert to an Input type.