pars
Trait Parse

pub trait Parse<I: Input>: Sized {
    type Parsed: Sized;
    type Error: Error<I>;

    // Required method
    fn parse<N>(&self, input: N) -> PResult<Self::Parsed, I, Self::Error>
       where N: IntoInput<Input = I>;

    // Provided methods
    fn map<F, R>(
        self,
        map_fn: F,
    ) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where F: Fn(Self::Parsed) -> R { ... }
    fn try_map<F, R, S>(
        self,
        map_fn: F,
    ) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where S: ErrorSeed<I, Self::Error>,
             F: Fn(Self::Parsed) -> Result<R, S> { ... }
    fn map_err<F, R>(
        self,
        map_err_fn: F,
    ) -> impl Parse<I, Parsed = Self::Parsed, Error = R>
       where F: Fn(Self::Error) -> R,
             R: Error<I> { ... }
    fn map_res<F, R, E>(
        self,
        map_res_fn: F,
    ) -> impl Parse<I, Parsed = R, Error = E>
       where F: Fn(Result<Self::Parsed, Self::Error>) -> Result<R, E>,
             E: Error<I> { ... }
    fn res_into<R, E>(self) -> impl Parse<I, Parsed = R, Error = E>
       where Self::Parsed: Into<R>,
             Self::Error: Into<E>,
             E: Error<I> { ... }
    fn ok_into<R>(self) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where Self::Parsed: Into<R> { ... }
    fn err_into<E>(self) -> impl Parse<I, Parsed = Self::Parsed, Error = E>
       where Self::Error: Into<E>,
             E: Error<I> { ... }
    fn res_try_into<R, E>(self) -> impl Parse<I, Parsed = R, Error = E>
       where Self::Parsed: TryInto<R>,
             Self::Error: Into<E>,
             E: Error<I> { ... }
    fn ok_try_into<R>(self) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where Self::Parsed: TryInto<R> { ... }
    fn with<F, T>(
        self,
        with_fn: F,
    ) -> impl Parse<I, Parsed = T, Error = Self::Error>
       where F: Fn() -> T { ... }
    fn with_default<T>(self) -> impl Parse<I, Parsed = T, Error = Self::Error>
       where T: Default { ... }
    fn with_value<T>(
        self,
        value: T,
    ) -> impl Parse<I, Parsed = T, Error = Self::Error>
       where T: Clone { ... }
    fn flat_map<F, P>(
        self,
        combinator: F,
    ) -> impl Parse<I, Parsed = P::Parsed, Error = Self::Error>
       where F: Fn(Self::Parsed) -> P,
             P: Parse<I, Error = Self::Error> { ... }
    fn try_flat_map<F, P, S>(
        self,
        combinator: F,
    ) -> impl Parse<I, Parsed = P::Parsed, Error = Self::Error>
       where S: ErrorSeed<I, Self::Error>,
             F: Fn(Self::Parsed) -> Result<P, S>,
             P: Parse<I, Error = Self::Error> { ... }
    fn or_value(
        self,
        value: Self::Parsed,
    ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
       where Self::Parsed: Clone { ... }
    fn or_default(
        self,
    ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
       where Self::Parsed: Default { ... }
    fn or_else<F>(
        self,
        or_else_fn: F,
    ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
       where F: Fn() -> Self::Parsed { ... }
    fn many0<F, R>(
        self,
        collect_fn: F,
    ) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where F: for<'a> Fn(Many0Iter<'a, Self, I>) -> R { ... }
    fn try_many0<F, R, S>(
        self,
        collect_fn: F,
    ) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where S: ErrorSeed<I, Self::Error>,
             F: for<'a> Fn(Many0Iter<'a, Self, I>) -> Result<R, S> { ... }
    fn collect_many0<R>(self) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where R: FromIterator<Self::Parsed> { ... }
    fn many1<F, R>(
        self,
        collect_fn: F,
    ) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where F: for<'a> Fn(Many1Iter<'a, Self, I>) -> R { ... }
    fn try_many1<F, R, S>(
        self,
        collect_fn: F,
    ) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where S: ErrorSeed<I, Self::Error>,
             F: for<'a> Fn(Many1Iter<'a, Self, I>) -> Result<R, S> { ... }
    fn collect_many1<R>(self) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where R: FromIterator<Self::Parsed> { ... }
    fn repeated<F, R>(
        self,
        count: usize,
        collect_fn: F,
    ) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where F: for<'a> Fn(RepeatedIter<'a, Self, I>) -> R { ... }
    fn try_repeated<F, R, S>(
        self,
        count: usize,
        collect_fn: F,
    ) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where S: ErrorSeed<I, Self::Error>,
             F: for<'a> Fn(RepeatedIter<'a, Self, I>) -> Result<R, S> { ... }
    fn collect_repeated<R>(
        self,
        count: usize,
    ) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where R: FromIterator<Self::Parsed> { ... }
    fn many_until<P, F, R>(
        self,
        sentinel: P,
        collect_fn: F,
    ) -> impl Parse<I, Parsed = (R, P::Parsed), Error = Self::Error>
       where P: Parse<I, Error = Self::Error>,
             F: for<'a> Fn(ManyUntilIter<'a, Self, P, I>) -> R { ... }
    fn try_many_until<P, F, R, S>(
        self,
        sentinel: P,
        collect_fn: F,
    ) -> impl Parse<I, Parsed = (R, P::Parsed), Error = Self::Error>
       where S: ErrorSeed<I, Self::Error>,
             P: Parse<I, Error = Self::Error>,
             F: for<'a> Fn(ManyUntilIter<'a, Self, P, I>) -> Result<R, S> { ... }
    fn collect_many_until<P, R>(
        self,
        sentinel: P,
    ) -> impl Parse<I, Parsed = (R, P::Parsed), Error = Self::Error>
       where P: Parse<I, Error = Self::Error>,
             R: FromIterator<Self::Parsed> { ... }
    fn many<N, F, R>(
        self,
        range: N,
        collect_fn: F,
    ) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where N: RangeBounds<usize>,
             F: for<'a> Fn(ManyIter<'a, Self, I>) -> R { ... }
    fn try_many<N, F, R, S>(
        self,
        range: N,
        collect_fn: F,
    ) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where N: RangeBounds<usize>,
             S: ErrorSeed<I, Self::Error>,
             F: for<'a> Fn(ManyIter<'a, Self, I>) -> Result<R, S> { ... }
    fn collect_many<N, R>(
        self,
        range: N,
    ) -> impl Parse<I, Parsed = R, Error = Self::Error>
       where N: RangeBounds<usize>,
             R: FromIterator<Self::Parsed> { ... }
    fn array<const LEN: usize>(
        self,
    ) -> impl Parse<I, Parsed = [Self::Parsed; LEN], Error = Self::Error> { ... }
    fn complete(
        self,
    ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error> { ... }
    fn spanned(
        self,
    ) -> impl Parse<I, Parsed = (Self::Parsed, Span<I>), Error = Self::Error> { ... }
    fn spanned_into<S>(
        self,
    ) -> impl Parse<I, Parsed = (Self::Parsed, S), Error = Self::Error>
       where Span<I>: Into<S> { ... }
    fn recognize(self) -> impl Parse<I, Parsed = Span<I>, Error = Self::Error> { ... }
    fn cond(
        self,
        condition: bool,
    ) -> impl Parse<I, Parsed = Option<Self::Parsed>, Error = Self::Error> { ... }
    fn verify<F>(
        self,
        verify_fn: F,
    ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
       where F: Fn(&Self::Parsed) -> bool { ... }
    fn not(self) -> impl Parse<I, Parsed = (), Error = Self::Error> { ... }
    fn opt(
        self,
    ) -> impl Parse<I, Parsed = Option<Self::Parsed>, Error = Self::Error> { ... }
    fn peek(self) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error> { ... }
    fn parse_as_ref<'a>(
        &'a self,
    ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error> + 'a
       where I: 'a { ... }
    fn then<P>(
        self,
        other: P,
    ) -> impl Parse<I, Parsed = (Self::Parsed, P::Parsed), Error = Self::Error>
       where P: Parse<I, Error = Self::Error> { ... }
    fn or<P>(
        self,
        other: P,
    ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
       where P: Parse<I, Parsed = Self::Parsed, Error = Self::Error> { ... }

}

Trait implemented by all parsers.

The Parse trait is where the logic of parsing is implemented. However, most pars users will not implement Parse directly. See the crate level documentation for examples of user defined parsers.

Required Associated Types

type Parsed: Sized

The value type that is produced by the parser on success.

type Error: Error<I>

The error type that is produced by the parser on failure.

Required Methods

fn parse<N>(&self, input: N) -> PResult<Self::Parsed, I, Self::Error>
where N: IntoInput<Input = I>,

Parses the provided input.

Provided Methods

fn map<F, R>(self, map_fn: F) -> impl Parse<I, Parsed = R, Error = Self::Error>
where F: Fn(Self::Parsed) -> R,

fn try_map<F, R, S>( self, map_fn: F, ) -> impl Parse<I, Parsed = R, Error = Self::Error>
where S: ErrorSeed<I, Self::Error>, F: Fn(Self::Parsed) -> Result<R, S>,

fn map_err<F, R>( self, map_err_fn: F, ) -> impl Parse<I, Parsed = Self::Parsed, Error = R>
where F: Fn(Self::Error) -> R, R: Error<I>,

fn map_res<F, R, E>(self, map_res_fn: F) -> impl Parse<I, Parsed = R, Error = E>
where F: Fn(Result<Self::Parsed, Self::Error>) -> Result<R, E>, E: Error<I>,

fn res_into<R, E>(self) -> impl Parse<I, Parsed = R, Error = E>
where Self::Parsed: Into<R>, Self::Error: Into<E>, E: Error<I>,

fn ok_into<R>(self) -> impl Parse<I, Parsed = R, Error = Self::Error>
where Self::Parsed: Into<R>,

fn err_into<E>(self) -> impl Parse<I, Parsed = Self::Parsed, Error = E>
where Self::Error: Into<E>, E: Error<I>,

fn res_try_into<R, E>(self) -> impl Parse<I, Parsed = R, Error = E>
where Self::Parsed: TryInto<R>, Self::Error: Into<E>, E: Error<I>,

fn ok_try_into<R>(self) -> impl Parse<I, Parsed = R, Error = Self::Error>
where Self::Parsed: TryInto<R>,

fn with<F, T>( self, with_fn: F, ) -> impl Parse<I, Parsed = T, Error = Self::Error>
where F: Fn() -> T,

fn with_default<T>(self) -> impl Parse<I, Parsed = T, Error = Self::Error>
where T: Default,

fn with_value<T>( self, value: T, ) -> impl Parse<I, Parsed = T, Error = Self::Error>
where T: Clone,

fn flat_map<F, P>( self, combinator: F, ) -> impl Parse<I, Parsed = P::Parsed, Error = Self::Error>
where F: Fn(Self::Parsed) -> P, P: Parse<I, Error = Self::Error>,

fn try_flat_map<F, P, S>( self, combinator: F, ) -> impl Parse<I, Parsed = P::Parsed, Error = Self::Error>
where S: ErrorSeed<I, Self::Error>, F: Fn(Self::Parsed) -> Result<P, S>, P: Parse<I, Error = Self::Error>,

fn or_value( self, value: Self::Parsed, ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
where Self::Parsed: Clone,

fn or_default(self) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
where Self::Parsed: Default,

fn or_else<F>( self, or_else_fn: F, ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
where F: Fn() -> Self::Parsed,

fn many0<F, R>( self, collect_fn: F, ) -> impl Parse<I, Parsed = R, Error = Self::Error>
where F: for<'a> Fn(Many0Iter<'a, Self, I>) -> R,

fn try_many0<F, R, S>( self, collect_fn: F, ) -> impl Parse<I, Parsed = R, Error = Self::Error>
where S: ErrorSeed<I, Self::Error>, F: for<'a> Fn(Many0Iter<'a, Self, I>) -> Result<R, S>,

fn collect_many0<R>(self) -> impl Parse<I, Parsed = R, Error = Self::Error>
where R: FromIterator<Self::Parsed>,

fn many1<F, R>( self, collect_fn: F, ) -> impl Parse<I, Parsed = R, Error = Self::Error>
where F: for<'a> Fn(Many1Iter<'a, Self, I>) -> R,

fn try_many1<F, R, S>( self, collect_fn: F, ) -> impl Parse<I, Parsed = R, Error = Self::Error>
where S: ErrorSeed<I, Self::Error>, F: for<'a> Fn(Many1Iter<'a, Self, I>) -> Result<R, S>,

fn collect_many1<R>(self) -> impl Parse<I, Parsed = R, Error = Self::Error>
where R: FromIterator<Self::Parsed>,

fn repeated<F, R>( self, count: usize, collect_fn: F, ) -> impl Parse<I, Parsed = R, Error = Self::Error>
where F: for<'a> Fn(RepeatedIter<'a, Self, I>) -> R,

fn try_repeated<F, R, S>( self, count: usize, collect_fn: F, ) -> impl Parse<I, Parsed = R, Error = Self::Error>
where S: ErrorSeed<I, Self::Error>, F: for<'a> Fn(RepeatedIter<'a, Self, I>) -> Result<R, S>,

fn collect_repeated<R>( self, count: usize, ) -> impl Parse<I, Parsed = R, Error = Self::Error>
where R: FromIterator<Self::Parsed>,

fn many_until<P, F, R>( self, sentinel: P, collect_fn: F, ) -> impl Parse<I, Parsed = (R, P::Parsed), Error = Self::Error>
where P: Parse<I, Error = Self::Error>, F: for<'a> Fn(ManyUntilIter<'a, Self, P, I>) -> R,

fn try_many_until<P, F, R, S>( self, sentinel: P, collect_fn: F, ) -> impl Parse<I, Parsed = (R, P::Parsed), Error = Self::Error>
where S: ErrorSeed<I, Self::Error>, P: Parse<I, Error = Self::Error>, F: for<'a> Fn(ManyUntilIter<'a, Self, P, I>) -> Result<R, S>,

fn collect_many_until<P, R>( self, sentinel: P, ) -> impl Parse<I, Parsed = (R, P::Parsed), Error = Self::Error>
where P: Parse<I, Error = Self::Error>, R: FromIterator<Self::Parsed>,

fn many<N, F, R>( self, range: N, collect_fn: F, ) -> impl Parse<I, Parsed = R, Error = Self::Error>
where N: RangeBounds<usize>, F: for<'a> Fn(ManyIter<'a, Self, I>) -> R,

fn try_many<N, F, R, S>( self, range: N, collect_fn: F, ) -> impl Parse<I, Parsed = R, Error = Self::Error>
where N: RangeBounds<usize>, S: ErrorSeed<I, Self::Error>, F: for<'a> Fn(ManyIter<'a, Self, I>) -> Result<R, S>,

fn collect_many<N, R>( self, range: N, ) -> impl Parse<I, Parsed = R, Error = Self::Error>
where N: RangeBounds<usize>, R: FromIterator<Self::Parsed>,

fn array<const LEN: usize>( self, ) -> impl Parse<I, Parsed = [Self::Parsed; LEN], Error = Self::Error>

fn complete(self) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>

fn spanned( self, ) -> impl Parse<I, Parsed = (Self::Parsed, Span<I>), Error = Self::Error>

fn spanned_into<S>( self, ) -> impl Parse<I, Parsed = (Self::Parsed, S), Error = Self::Error>
where Span<I>: Into<S>,

fn recognize(self) -> impl Parse<I, Parsed = Span<I>, Error = Self::Error>

fn cond( self, condition: bool, ) -> impl Parse<I, Parsed = Option<Self::Parsed>, Error = Self::Error>

fn verify<F>( self, verify_fn: F, ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
where F: Fn(&Self::Parsed) -> bool,

fn not(self) -> impl Parse<I, Parsed = (), Error = Self::Error>

fn opt( self, ) -> impl Parse<I, Parsed = Option<Self::Parsed>, Error = Self::Error>

fn peek(self) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>

fn parse_as_ref<'a>( &'a self, ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error> + 'a
where I: 'a,

fn then<P>( self, other: P, ) -> impl Parse<I, Parsed = (Self::Parsed, P::Parsed), Error = Self::Error>
where P: Parse<I, Error = Self::Error>,

fn or<P>( self, other: P, ) -> impl Parse<I, Parsed = Self::Parsed, Error = Self::Error>
where P: Parse<I, Parsed = Self::Parsed, Error = Self::Error>,

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.